==> scripts/2.1_exploratory_analyses.Rmd <==
---
title: "Exploratory Analyses"
author: "Joe Celebrezze"
date: "2023-03-30"
output: html_document
---

This script includes a bunch of exploratory data analyses and associated figures; it has not been updated or cleaned in a little while, but it still runs okay and exhibits things we considered when running EDA and formulating an understanding of the dataset

# Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(here)
here = here::here
source(here("scripts", "source_code.R"))
```

# Summary Boxplots
## Plant Traits
```{r}
traits.df.long <- flamm.df %>% 
  select(date, species, plant, rep, leaf_lfm, stem_lfm, lfm, leaf_area, SLA, LMA,
         thickness, branch_volume, sample_wt,
         branching, mpa, stem_mass_ratio, leaf_mass_ratio) %>% 
  pivot_longer(!c(date, species, plant, rep), names_to = 'trait', values_to = 'value')

ggplot(traits.df.long, aes(x = species, y = value, color = species)) +
  geom_boxplot() +
  labs(x = "Species", y = "Trait Value") +
  facet_wrap(~trait, ncol = 3, scales = 'free_y') +
  scale_color_manual(values = c('#ADB6B6B2', '#1B1919B2', '#42B540B2', '#ED0000B2', '#AD002AB2', '#0099B4B2', '#925E9FB2', '#925E9FB2', '#FDAF91B2', '#00468BB2')) +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 20),
        axis.text = element_text(size = 18),
        strip.text = element_text(face = 'bold', size = 20),
        axis.text.x = element_text(angle = 18))

ggsave(here('figures', 'EDA', 'plant.traits.boxplot.jpg'), height = 20, width = 26)
```

## Flammability Metrics
```{r}
flam_labels <- c('Flame Height', 'Flame Duration', 'Temperature Change', 'Heat Flux Change', 'Post-Flame Glow', '% Ignited')
names(flam_labels) <- c('fh', 'fd', 'temp_change', 'heat_flux_change', 'pfg', 'prop_ig')

flamm.df.long <- flamm.df %>% 
  select(date, species, plant, rep, mpa, branching, branch_height, branch_length, branch_width, leaf_lfm, leaf_lfm, stem_lfm, lfm, LMA, SLA, thickness, sample_wt, fh, fd, tti, pfg, max_temp, temp_change, max_heat_flux_loessCH7, max_heat_flux_loessCH8, dw_flam_sample, ww_flam_sample, branch_volume, heat_flux_change, stem_sav, leaf_sav, dmc, leaf_dmc, stem_dmc, dw_sppdev) %>%  # took out  gd, gti, ttfg for now
  pivot_longer(!c(date, species, plant, rep, mpa, branching, branch_height, branch_length, branch_width, leaf_lfm, stem_lfm, lfm, LMA, SLA, thickness, sample_wt, dw_flam_sample, ww_flam_sample, branch_volume, stem_sav, leaf_sav, dmc, leaf_dmc, stem_dmc, dw_sppdev), names_to = 'trait', values_to = 'value') %>% 
  mutate(species = factor(species, levels = c('SALAPI', 'SALLEU', 'CEAGRI', 'ARTCAL', 'QUEAGR', 'MALLAU', 'EUCGLO', 'ERIKAR', 'HETARB', 'IRIDOU', 'ARCDEN'))) #note that EUGL needs to change to EUCGLO

flamm.df.long %>% 
  filter(trait %in% c('fd', 'fh', 'pfg', 'heat_flux_change', 'temp_change')) %>% 
ggplot(aes(x = species, y = value, color = species)) +
  geom_boxplot() +
  labs(x = "Species", y = "Trait Value") +
  facet_wrap(~trait, ncol = 3, scales = 'free_y', labeller = labeller(trait = flam_labels)) +
  scale_color_manual(values = c('#ADB6B6B2', '#1B1919B2', '#42B540B2', '#ED0000B2', '#AD002AB2', '#0099B4B2', '#925E9FB2', '#925E9FB2', '#FDAF91B2', '#00468BB2')) +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        strip.text = element_text(face = 'bold', size = 16),
        axis.text.x = element_text(angle = 20, size = 12))

ggsave(here('figures', 'EDA', 'flamm.metrics.boxplot.jpg'), height = 12, width = 20)
```

## Normalized Flamm. Metrics
```{r}
flamm.df.long %>% 
  mutate(norm_value = value/as.numeric(sample_wt)) %>% 
ggplot(aes(x = species, y = norm_value, color = species)) +
  geom_boxplot() +
  labs(x = "Species", y = "Normalized Trait Value") +
  facet_wrap(~trait, ncol = 4, scales = 'free_y') +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        strip.text = element_text(face = 'bold', size = 16),
        axis.text.x = element_text(angle = 20, size = 12))

flamm.df.long %>% 
  filter(trait %in% c('fd', 'fh', 'pfg')) %>% 
  mutate(norm_value = value/as.numeric(sample_wt)) %>% 
ggplot(aes(x = species, y = norm_value, color = species)) +
  geom_boxplot() +
  labs(x = "Species", y = "Normalized Trait Value") +
  facet_wrap(~trait, ncol = 4, scales = 'free_y') +
  scale_color_manual(values = c('#ADB6B6B2', '#1B1919B2', '#42B540B2', '#ED0000B2', '#AD002AB2', '#0099B4B2', '#925E9FB2', '#925E9FB2', '#FDAF91B2', '#00468BB2')) +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        strip.text = element_text(face = 'bold', size = 16),
        axis.text.x = element_text(angle = 20, size = 12))

ggsave(here('figures', 'EDA', 'normalized.flamm.metrics.boxplot.jpg'), height = 12, width = 20)
```

# Scatterplots
```{r}
flamm.df %>% 
  pivot_longer(cols = c('fh', 'fd', 'temp_change', 'heat_flux_change'),
               names_to = 'flam_metric', values_to = 'flam_val') %>% 
  pivot_longer(cols = c('lfm', 'dmc', 'mpa', 'branching', 'stem_sav', 'leaf_mass_ratio',
                        'leaf_sav', 'LMA', 'thickness'),
               names_to = 'predictor', values_to = 'pred_val') %>% 
  ggplot(aes(x = pred_val, y = flam_val)) +
    geom_point(alpha = 0.4) +
    geom_smooth(method = 'lm', color = 'black', fill = 'gray80') +
    stat_cor()+ 
    #stat_regline_equation() +
    facet_grid(flam_metric~predictor, scales = 'free', switch = 'both') +
    theme_bw() +
    theme(strip.background = element_blank(),
          strip.placement = 'outside',
          axis.title = element_blank())
```

## Function
```{r}
flamm.df.long.scatter <- flamm.df.long %>% 
  dplyr::filter(trait %in% c('fd', 'fh', 'temp_change', 'heat_flux_change', 'pfg'))

exploratory_scatterplot <- function(expl.var, x.lab){
flamm.df.long.scatter %>% 
  ggplot(aes(x = expl.var, y = value, color = species)) +
    geom_point(alpha = 0.6, size = 1.5) +
  geom_smooth(method = 'lm', linewidth = 1, alpha = 0.8, se = F) +
  labs(x = x.lab, y = 'Flamm. Metric Value', color = 'Species')  +
  facet_wrap(~trait, ncol = 3, scales = 'free_y', labeller = labeller(trait = flam_labels)) +
  scale_color_manual(values = c('#ADB6B6B2', '#1B1919B2', '#42B540B2', '#ED0000B2', '#AD002AB2', '#0099B4B2', '#925E9FB2', '#925E9FB2', '#FDAF91B2', '#00468BB2')) +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        legend.title = element_text(face = 'bold', size = 16),
        legend.text = element_text(size = 14),
        strip.text = element_text(face = 'bold', size = 16))
}
```

## Vs. Hydration
Water Potential
```{r}
exploratory_scatterplot(flamm.df.long.scatter$mpa, 'Water Potential (MPa)')

ggsave(here('figures', 'EDA', 'flamm.metrics.vs.mpa.jpg'), height = 12, width = 20)
```

LFM
```{r}
exploratory_scatterplot(flamm.df.long.scatter$leaf_lfm, 'Leaf Live Fuel Moisture (%)')
exploratory_scatterplot(flamm.df.long.scatter$stem_lfm, 'Stem Live Fuel Moisture (%)')
exploratory_scatterplot(flamm.df.long.scatter$lfm, 'Live Fuel Moisture (%)')
ggsave(here('figures', 'EDA', 'flamm.metrics.vs.lfm.jpg'), height = 12, width = 20)
```

## Vs. Sample Weight
and DW, WW
```{r}
exploratory_scatterplot(flamm.df.long.scatter$sample_wt, 'Sample Weight (g)')
ggsave(here('figures', 'EDA', 'flamm.metrics.vs.sample.wt.jpg'), height = 12, width = 20)

exploratory_scatterplot(flamm.df.long.scatter$dw_flam_sample, 'Dry Weight (g)')
ggsave(here('figures', 'EDA', 'flamm.metrics.vs.dw.jpg'), height = 12, width = 20)

exploratory_scatterplot(flamm.df.long.scatter$ww_flam_sample, 'Wet Weight (g)')
ggsave(here('figures', 'EDA', 'flamm.metrics.vs.ww.jpg'), height = 12, width = 20)

exploratory_scatterplot(flamm.df.long.scatter$dw_sppdev, 'Dry Weight (Dev. from Sp. Mean)')
```

## Vs. Leaf Metrics
```{r}
exploratory_scatterplot(flamm.df.long.scatter$LMA, 'Leaf Mass per Area')
exploratory_scatterplot(flamm.df.long.scatter$SLA, 'Specific Leaf Area')
exploratory_scatterplot(flamm.df.long.scatter$thickness, 'Leaf Thickness')
exploratory_scatterplot(flamm.df.long.scatter$leaf_dmc, 'Leaf Dry Matter Content')
exploratory_scatterplot(flamm.df.long.scatter$leaf_sav, 'Leaf SA:V')
```

## Vs. Stem Metrics
```{r}
exploratory_scatterplot(flamm.df.long.scatter$stem_dmc, 'Stem Dry Matter Content')
exploratory_scatterplot(flamm.df.long.scatter$stem_sav, 'Stem SA:V')
exploratory_scatterplot(flamm.df.long.scatter$dmc, 'Dry Matter Content')
```

# Proportion Ignited
```{r}
flamm.df %>% 
  group_by(species) %>%
  dplyr::summarise(n = n(), prop_ig = prop_ig) %>% 
  mutate(species = factor(species, levels = c('SALAPI', 'SALLEU', 'CEAGRI', 'ARTCAL', 'MALLAU', 'ERIKAR', 'HETARB', 'IRIDOU', 'ARCDEN'))) %>% 
  ggplot(aes(x = species, y = prop_ig, color = species)) +
    scale_y_continuous(limits = c(0,1)) +
    geom_point(size = 10) +
    scale_color_manual(values = c('#ADB6B6B2', '#1B1919B2', '#42B540B2', '#ED0000B2', '#AD002AB2', '#0099B4B2', '#925E9FB2', '#FDAF91B2', '#00468BB2')) +
    geom_text(aes(label = paste0('N = ', n), x = species, y = prop_ig + 0.1)) +
    labs(x = ' ', y = 'Proportion Ignited') +
    theme_bw() + 
    theme(legend.position = 'none',
          axis.title.y = element_text(face = 'bold', size = 15),
          axis.text.y = element_text(size = 14),
          axis.title.x = element_blank(),
          axis.text.x = element_text(size = 12, angle = 15, vjust = 0.9, hjust=0.75))

ggsave(here('figures', 'main_figures', 'proportion.ignited.jpg'), height = 6, width = 8)
```

# PCAs
## Flamm Metrics
```{r}
# selecting metrics of interest
flamm_pca_df <- flamm.df %>% 
  select(fd, fh, prop_ig, heat_flux_change, temp_change, pfg, species, ignition) # removed tti because that's mostly tied to timing of manual ignition

# looking at correlation and distribution
pairs.panels(flamm_pca_df)

# prepping dataset for PCA
flamm_pca_df <- flamm_pca_df %>% 
  na.omit() %>%   # Getting rid of NAs, which currently drops 113 samples :( -- need to find a way to limit
  filter(temp_change > -100) # Getting rid of -Inf value

# running PCA
flamm_pca <- flamm_pca_df %>% 
  select(-species, -ignition) %>% 
  prcomp(center = TRUE, scale = TRUE) 

# first glance at results
summary(flamm_pca)
plot(flamm_pca, type = 'lines') # first 2 explain 51.86% variance
print(flamm_pca) # PC1: flame height, flame duration, max temp and max heat flux (combustibility), PC2: time to ignition, proportion ignited, post-flame glow (ignitability and consumability)

# visualizing results
ggbiplot(flamm_pca,
         groups = flamm_pca_df$species,
         ellipse = TRUE, varname.size = 5, alpha = 0.5) +
  scale_color_lancet() +
  theme_bw()
# eigenvalue table
tab_pca(flamm_pca, rotation = 'varimax')
```

## Leaf Traits
```{r}
# selecting metrics of interest
traits_pca_df <- flamm.df %>% 
  select(leaf_mass_ratio, leaf_lfm, stem_lfm, lfm, thickness, mpa, SLA, dw_sppdev, leaf_sav, leaf_dmc, dmc, stem_dmc, stem_sav, branch_volume, species, ignition)

# looking at correlation and distribution
pairs.panels(traits_pca_df) # from psych package

# prepping dataset for PCA
traits_pca_df <- traits_pca_df %>% 
  na.omit()   # Getting rid of NAs, which currently drops 36 samples

# running PCA
traits_pca <- traits_pca_df %>% 
  select(-species, -ignition) %>% 
  prcomp(center = TRUE, scale = TRUE)  #prcomp from stats package

# first glance at results
summary(traits_pca)
plot(traits_pca, type = 'lines') # first 2 explain 47.99% variance
print(traits_pca) # PC1: stem LFM, LFM, leaf mass ratio, PC2: SLA, lfm, leaf LFM, PC3: LFM, MPa

# visualizing results, from package ggbiplot
ggbiplot(traits_pca,
         groups = traits_pca_df$species,
         ellipse = TRUE, varname.size = 5, alpha = 0.5) +
  scale_color_lancet() +
  theme_bw()

# eigenvalue table
tab_pca(traits_pca, rotation = 'varimax') # From package sjPlot
```

# k-means Clustering
## Flamm Metrics
NOTE: Ignition method has significant effect on clustering
```{r}
# so it's replicable
set.seed(16)

# prepping dataset
id <- c(1:nrow(flamm_pca_df))
flamm_kmeans_df <- cbind(flamm_pca_df, id) %>% 
  unite('species_id', c(species, id), remove = F)  #%>% 
  #filter(ignition == 'M')

rownames(flamm_kmeans_df) <- flamm_kmeans_df$species_id

flamm_kmeans_scaled_df <- flamm_kmeans_df %>% 
  #filter(ignition == 'M') %>% 
  select(-c(species_id, species, id, prop_ig, ignition)) %>%  #removing prop_ig for now as it's directly tied to species
  scale() %>% 
  as.data.frame()

# optimal number of clusters
# 2 ways to determine optimal cluster 
# 1. optimal cluster at the elbow
fviz_nbclust(flamm_kmeans_scaled_df, kmeans, method = "wss") #k = 2 or k = 3
# 2. silhouette method
fviz_nbclust(flamm_kmeans_scaled_df, kmeans, method = "silhouette") #k = 2, 7, or 10

# run k-means clustering
#perform k-means clustering with k = 2 clusters
km <- kmeans(flamm_kmeans_scaled_df, centers = 2)
#plot results of final k-means model
fviz_cluster(km, data = flamm_kmeans_scaled_df)
#add cluster assigment to original data
flamm_kmeans_df <- cbind(flamm_kmeans_df, cluster = km$cluster)

#visualizing how cluster assignment tracks species
flamm_kmeans_df %>% 
  ggplot(aes(x = species, y = cluster, color = as.factor(cluster))) +
    geom_jitter(width = 0.1, height = 0.1) +
    scale_y_continuous(breaks = c(1:max(flamm_kmeans_df$cluster))) +
    theme_bw() +
    labs(x = ' ', y = 'Cluster #', color = 'Cluster #', title = 'K-Means Clustering: Flam. Metrics')

flamm_kmeans_df %>% 
  ggplot(aes(x = cluster, fill = species)) +
    geom_bar() +
    theme_bw() +
    labs(x = 'Cluster #', fill = 'Species', y = 'Count', title = 'K-Means Clustering: Flam. Metrics')

flamm_kmeans_df %>% 
  pivot_longer(cols = c(fd, fh, prop_ig, heat_flux_change, temp_change, pfg), names_to = 'var', values_to = 'val') %>% 
  ggplot(aes(x = as.factor(cluster), y = val, color = as.factor(cluster))) +
    geom_boxplot(outlier.shape = NA, alpha= 0.8) +
    geom_jitter(width = 0.2, alpha = 0.2) +
    facet_wrap(~var, scales = 'free', labeller = labeller(var = flam_labels)) +
    labs(x = 'K-Means Cluster', y = 'Flamm. Metric Value') +
    scale_color_manual(values = c('goldenrod4', 'gold3', 'firebrick')) +
    theme_bw() +
    theme(legend.position = 'none',
          axis.title = element_text(face = 'bold', size = 17),
          axis.text = element_text(size = 15),
          strip.text = element_text(face = 'bold', size = 15))

flamm_kmeans_df %>% 
  group_by(cluster, species) %>% 
  tally() %>%  
  pivot_wider(names_from = cluster, values_from = n) %>% 
  mutate(`1` = ifelse(is.na(`1`), 0, `1`),
         `2` = ifelse(is.na(`2`), 0, `2`)) %>% 
  mutate(kmeans.flam.index = (`1`)/(`2` + `1`)) %>% 
  ggplot(aes(x = species, y = kmeans.flam.index, color = species)) +
    geom_point(size = 7) +
    labs(x = 'Species', y = 'K-Means Clustering: Flammability Index') +
    theme_bw() +
    theme(legend.position = 'none',
          axis.title = element_text(face = 'bold', size = 14),
          axis.text = element_text(size = 13),
          axis.text.x = element_text(angle = 15))

flamm_kmeans_df %>% 
  group_by(ignition, cluster) %>% 
  tally()

ignition.vs.cluster <- lm(cluster ~ ignition, data = flamm_kmeans_df)
anova(ignition.vs.cluster)
summary(ignition.vs.cluster)
```

## Leaf Traits
```{r}
# prepping dataset
id <- c(1:nrow(traits_pca_df))
traits_kmeans_df <- cbind(traits_pca_df, id) %>% 
  unite('species_id', c(species, id), remove = F)

rownames(traits_kmeans_df) <- traits_kmeans_df$species_id

traits_kmeans_scaled_df <- traits_kmeans_df %>% 
  select(-c(species_id, species, id, ignition)) %>% 
  scale() %>% 
  as.data.frame()

# optimal number of clusters
# 2 ways to determine optimal cluster 
# 1. optimal cluster at the elbow
fviz_nbclust(traits_kmeans_scaled_df, kmeans, method = "wss") #k = 3 or 8
# 2. silhouette method
fviz_nbclust(traits_kmeans_scaled_df, kmeans, method = "silhouette") #k = 3, 8 or 10

# run k-means clustering
#perform k-means clustering with k = 3 clusters
km <- kmeans(traits_kmeans_scaled_df, centers = 3)
#plot results of final k-means model
fviz_cluster(km, data = traits_kmeans_scaled_df)
#add cluster assigment to original data
traits_kmeans_df <- cbind(traits_kmeans_df, cluster = km$cluster)

#visualizing how cluster assignment tracks species
traits_kmeans_df %>% 
  ggplot(aes(x = species, y = cluster, color = as.factor(cluster))) +
    geom_jitter(width = 0.1, height = 0.1) +
    scale_y_continuous(breaks = c(1:max(traits_kmeans_df$cluster))) +
    theme_bw() +
    labs(x = ' ', y = 'Cluster #', color = 'Cluster #', title = 'K-Means Clustering: Plant Traits')

traits_kmeans_df %>% 
  ggplot(aes(x = cluster, fill = species)) +
    geom_bar() +
    theme_bw() +
    labs(x = 'Cluster #', fill = 'Species', y = 'Count', title = 'K-Means Clustering: Plant Traits')
```

# Binomial Regressions
Looking at ignition vs. non-ignition/manual ignition
## Visualizations
```{r}
# prepping dataset
flamm_binomial_df <- flamm.df %>% 
  mutate(ignition = case_when(ignition == 'M' ~ 0,
                             ignition == '1' ~ 1,
                             ignition == '0' ~ 0)) %>% 
  filter(species %in% c('SALLEU', 'SALAPI', 'ARTCAL', 'CEAGRI', 'QUEAGR')) # focusing on species with at least 2 ignitions

# writing function
exploratory_binomial <- function(expl.var, x.lab){
  ggplot(data = flamm_binomial_df, aes(x=expl.var, y=ignition)) + 
  geom_point(alpha=.5) +
  facet_wrap(~species, scales = 'free_x') +
  stat_smooth(method="glm", se=FALSE, method.args = list(family=binomial)) +
  scale_y_continuous(breaks = c(0,1)) +
  theme_bw() +
  theme(strip.text = element_text(face = 'bold', size = 15),
        axis.title = element_text(face = 'bold', size = 16)) +
  labs(x = x.lab, y = "Ignition?", title = "Binomial Regression")}

# looking at a bunch of different explanatory variables
exploratory_binomial(flamm_binomial_df$mpa, 'Water Potential (MPa)')
exploratory_binomial(flamm_binomial_df$lfm, 'Live Fuel Moisture (%)')
exploratory_binomial(flamm_binomial_df$leaf_lfm, 'Leaf Live Fuel Moisture (%)')
exploratory_binomial(flamm_binomial_df$stem_lfm, 'Stem Live Fuel Moisture (%)')
exploratory_binomial(flamm_binomial_df$LMA, 'Leaf Mass Area')
exploratory_binomial(flamm_binomial_df$sample_wt, 'Sample Weight (g)')
exploratory_binomial(flamm_binomial_df$start_temp, 'Starting Temperature (C)')
```

## Significance
```{r}
# writing function
binomial_significance <- function(spp){
df <- flamm_binomial_df %>% 
  filter(species == spp)

m1 <- glm(ignition ~ lfm + sample_wt, data = df, family = "binomial")
print(paste0(spp, ': LFM P-value = ', summary(m1)$coefficients[2,4]))

m2 <- glm(ignition ~ mpa + sample_wt, data = df, family = "binomial")
print(paste0(spp,': MPa P-value = ', summary(m2)$coefficients[2,4]))

m3 <- glm(ignition ~ sample_wt, data = df, family = "binomial")
print(paste0(spp, ': Sample Wt. P-value = ', summary(m3)$coefficients[2,4]))

df2 <- df %>% 
  filter(start_temp > 0)
m4 <- glm(ignition ~ start_temp + sample_wt, data = df2, family = "binomial")
print(paste0(spp, ': Starting Temp. P-value = ', summary(m4)$coefficients[2,4]))

m5 <- glm(ignition ~ LMA + sample_wt, data = df, family = "binomial")
print(paste0(spp, ': LMA P-value = ', summary(m5)$coefficients[2,4]))
}

flamm.df %>% 
  filter(start_temp > 0) %>% 
  summarise(avg = mean(start_temp))

# Running on all species that have at least 2 ignitions
binomial_significance('SALLEU')
binomial_significance('SALAPI')
#binomial_significance('ARTCAL') # sample weight, MPa are significant; note that because LMA is screwy for ARTCAL, this was an error
binomial_significance('CEAGRI')
```

# Investigating Hot Plate/Thermocoupler Height
## Hot Plate Height
```{r}
# Does it affect ignition percentage?
flamm.df.hp.ht.prop.ig1 <- flamm.df %>% 
  group_by(hotplate_height, species) %>%
  dplyr::summarise(n = n()) %>% 
  ungroup() 
  
flamm.df.hp.ht.prop.ig2 <- flamm.df %>% 
  filter(ignition %in% c('M', '0')) %>% 
  group_by(hotplate_height, species) %>% 
  dplyr::summarise(manual_or_nonignitions = n()) %>% 
  ungroup() %>% 
  select(manual_or_nonignitions)

flamm.df.hp.ht.prop.ig <- cbind(flamm.df.hp.ht.prop.ig1, flamm.df.hp.ht.prop.ig2) %>% 
  mutate(prop_ig = (n-manual_or_nonignitions)/n)

rm(flamm.df.hp.ht.prop.ig1, flamm.df.hp.ht.prop.ig2) # Decluttering environment by removing temporary dataframes

flamm.df.hp.ht.prop.ig %>% 
  filter(n > 5) %>% 
  mutate(species = factor(species, levels = c('SALAPI', 'SALLEU', 'CEAGRI', 'ARTCAL', 'MALLAU', 'ERIKAR', 'HETARB', 'IRIDOU', 'ARCDEN'))) %>% 
  ggplot(aes(x = hotplate_height, y = prop_ig, color = species)) +
    scale_y_continuous(limits = c(0,1)) +
    scale_color_manual(values = c('#ADB6B6B2', '#1B1919B2', '#42B540B2', '#ED0000B2', '#AD002AB2', '#0099B4B2', '#925E9FB2', '#925E9FB2', '#FDAF91B2', '#00468BB2')) +
    geom_jitter(aes(size = n), width = 0.1, height = 0) +
    #geom_text(aes(label = n, x = hotplate_height, y = prop_ig + 0.1)) +
    labs(x = 'Hot Plate Height (cm)', y = 'Proportion Ignited') +
    theme_bw()
#############################################
# For natural ignitions, does it affect time to ignition?
flamm.df %>% 
  filter(ignition == '1') %>% 
  ggplot(aes(x = as.numeric(hotplate_height), y = tti)) +
    geom_point() +
    geom_smooth(color = 'black') +
    labs(x = 'Hot Plate Height (cm)', y = 'Time to Ignition') +
    theme_bw()
#############################################
# Does it affect max. heat flux, temp. metrics?
flamm.df %>% 
  ggplot(aes(x = as.numeric(hotplate_height), y = max_heat_flux_loessCH8)) +
    geom_point() +
    geom_smooth(color = 'black') +
    labs(x = 'Hot Plate Height (cm)', y = 'Max. Heat Flux (Loess-Smoothed, Side Sensor)') +
    theme_bw()

flamm.df %>% 
  ggplot(aes(x = as.numeric(hotplate_height), y = max_heat_flux_loessCH7)) +
    geom_point() +
    geom_smooth(color = 'black') +
    labs(x = 'Hot Plate Height (cm)', y = 'Max. Heat Flux (Loess-Smoothed, Top Sensor)') +
    theme_bw()

flamm.df %>% 
  ggplot(aes(x = as.numeric(hotplate_height), y = max_temp)) +
    geom_point() +
    geom_smooth(color = 'black') +
    labs(x = 'Hot Plate Height (cm)', y = 'Max. Temp.') +
    theme_bw()

flamm.df %>% 
  ggplot(aes(x = as.numeric(hotplate_height), y = start_temp)) +
    geom_point() +
    geom_smooth(color = 'black') +
    labs(x = 'Hot Plate Height (cm)', y = 'Starting Temp.') +
    theme_bw()
```

## Thermocoupler Height
```{r}
# Does it affect temp. metrics?
flamm.df %>% 
  ggplot(aes(x = as.factor(thermocoupler_height), y = temp_change)) +
    geom_jitter(width = 0.2) +
    geom_boxplot(alpha = 0.2, outlier.shape = NA) +
    labs(x = 'Thermocoupler Height (cm)', y = 'Temp. Change') +
    theme_bw()
t.test(temp_change ~ thermocoupler_height, data = flamm.df) # significant
t.test(max_temp ~ thermocoupler_height, data = flamm.df) # significant
t.test(heat_flux_change ~ thermocoupler_height, data = flamm.df) # not significant (as expected)

flamm.df %>% 
  ggplot(aes(x = as.factor(thermocoupler_height), y = start_temp)) +
    geom_jitter(width = 0.2) +
    geom_boxplot(alpha = 0.2, outlier.shape = NA) +
    labs(x = 'Thermocoupler Height (cm)', y = 'Starting Temp.') +
    theme_bw()
t.test(start_temp ~ thermocoupler_height, data = flamm.df)
```

# Sampling Date vs. Flam. Metrics
```{r}
flamm.df %>%   
  select(date, species, plant, rep, fh, fd, lfm, mpa, temp_change, heat_flux_change) %>%  # took out  gd, gti, ttfg for now
  pivot_longer(!c(date, species, plant, rep), names_to = 'flam_metric', values_to = 'value') %>% 
  ggplot(aes(x = date, y = value, color = species)) +
    geom_point(alpha = 0.6, size = 1.5) +
    labs(x = 'Date', y = 'Flamm. Metric Value', color = 'Species')  +
    facet_wrap(~flam_metric, ncol = 3, scales = 'free_y') +
    theme_bw() +
    theme(axis.title = element_text(face = 'bold', size = 16),
          axis.text.y = element_text(size = 14),
          axis.text.x = element_blank(),
          axis.ticks.x = element_blank(),
          legend.title = element_text(face = 'bold', size = 16),
          legend.text = element_text(size = 14),
          strip.text = element_text(face = 'bold', size = 16))
# Tough to parse species differences and date differences

spp <- c('ARTCAL', 'CEAGRI', 'ERIKAR', 'HETARB', 'MALLAU', 'ARCDEN', 'SALAPI', 'SALLEU')
plot.list <- list(NULL)
for(i in 1:length(spp)) {
  df <- flamm.df %>%  
    filter(species == spp[i]) %>% 
    select(date, species, plant, rep, fh, fd, lfm, mpa, temp_change, heat_flux_change) %>%  # took out  gd, gti, ttfg for now
    pivot_longer(!c(date, species, plant, rep), names_to = 'flam_metric', values_to = 'value')

  plot <- ggplot(aes(x = date, y = value), data = df) +
    geom_point(alpha = 0.6, size = 1.5) +
    labs(x = 'Date', y = 'Flamm. Metric Value', title = paste(spp[i]))  +
    facet_wrap(~flam_metric, ncol = 3, scales = 'free_y') +
    theme_bw() +
    theme(axis.title = element_text(face = 'bold', size = 16),
          axis.text.y = element_text(size = 14),
          axis.text.x = element_blank(),
          axis.ticks.x = element_blank())
  plot.list[[i]] <- plot
}
plot.list[[1]]
plot.list[[2]]
plot.list[[3]]
plot.list[[4]]
plot.list[[5]]
plot.list[[6]]
plot.list[[7]]
plot.list[[8]]
```

# Manual Ignitions
Looking into differences in TTI for manual ignitions and how that affects other flam. metrics
```{r}
flamm.df.manual <- flamm.df %>% 
  filter(ignition == 'M')

# First, interspecific differences in time to ignition
flamm.df.manual %>% 
  mutate(species = factor(species, levels = c('CEAGRI', 'ARTCAL', 'ARCDEN', 'SALAPI', 'MALLAU', 'SALLEU', 'ERIKAR', 'HETARB', 'QUEAGR'))) %>% 
  ggplot(aes(x = species, y = tti)) + 
    geom_jitter(width = 0.2, aes(color = species)) +
    geom_boxplot(alpha = 0.2, outlier.shape = NA) +
    labs(x = ' ', y = 'Time to Ignition') +
    theme(legend.position = 'none') +
    theme_bw()

# Now, comparing to other flam. metrics
### Prepping Dataset ###
flamm.df.manual.long <- flamm.df.manual %>% 
  select(date, species, plant, rep, fh, fd, tti, pfg, temp_change, heat_flux_change) %>%  # took out  gd, gti, ttfg for now
  pivot_longer(!c(date, species, plant, rep, tti), names_to = 'trait', values_to = 'value')
### Scatterplots ###
flamm.df.manual.long %>% 
  ggplot(aes(x = tti, y = value)) +
  geom_point(alpha = 0.6, size = 1.5) +
  geom_smooth(method = 'lm', linewidth = 1, alpha = 0.8, se = F) +
  labs(x = 'Time to Ignition', y = 'Flamm. Metric Value', title = 'Manual Ignitions Exploratory Analysis')  +
  facet_wrap(~trait, ncol = 3, scales = 'free_y') +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        legend.title = element_text(face = 'bold', size = 16),
        legend.text = element_text(size = 14),
        strip.text = element_text(face = 'bold', size = 16))
# ~~ by species ~~ #
flamm.df.manual.long %>% 
  ggplot(aes(x = tti, y = value, color = species)) +
  geom_point(alpha = 0.6, size = 1.5) +
  geom_smooth(method = 'lm', linewidth = 1, alpha = 0.8, se = F) +
  labs(x = 'Time to Ignition', y = 'Flamm. Metric Value', color = 'Species', title = 'Manual Ignitions Exploratory Analysis')  +
  facet_wrap(~trait, ncol = 4, scales = 'free_y') +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        legend.title = element_text(face = 'bold', size = 16),
        legend.text = element_text(size = 14),
        strip.text = element_text(face = 'bold', size = 16))
### Linear Regressions ###
fd.tti.lm <- lm(fd ~ tti, data = flamm.df.manual)
summary(fd.tti.lm) # n.s.

fh.tti.lm <- lm(fh ~ tti, data = flamm.df.manual)
summary(fh.tti.lm) # n.s.

pfg.tti.lm <- lm(pfg ~ tti, data = flamm.df.manual)
summary(pfg.tti.lm) # n.s.

temp_change.tti.lm <- lm(temp_change ~ tti, data = flamm.df.manual)
summary(temp_change.tti.lm) # significant!

hf_change.tti.lm <- lm(heat_flux_change ~ tti, data = flamm.df.manual)
summary(hf_change.tti.lm)
# ~~ by species ~~ #
fdxsp.tti.lm <- lm(fd ~ tti*species, data = flamm.df.manual)
summary(fdxsp.tti.lm) # n.s.
fhxsp.tti.lm <- lm(fh ~ tti*species, data = flamm.df.manual)
summary(fhxsp.tti.lm) # n.s.
pfgxsp.tti.lm <- lm(pfg ~ tti*species, data = flamm.df.manual)
summary(pfgxsp.tti.lm) # tti x MALLAU, tti x CEAGRI sig.
tempxsp.tti.lm <- lm(temp_change ~ tti*species, data = flamm.df.manual)
summary(tempxsp.tti.lm) # tti x HETARB sig.
hfxsp.tti.lm <- lm(heat_flux_change ~ tti*species, data = flamm.df.manual)
summary(hfxsp.tti.lm) # n.s.
```

# Sample Weight Exploration
First, independent of species, look into linear and nonlinear relationships with flam. metrics (for now, FD, FH, PFG) and hydration

## Linear Relationships
```{r}
## FD ## - SIGNIFICANT
fd.sw.lm <- lm(scale(fd) ~ scale(sample_wt), data = flamm.df)
summary(fd.sw.lm)
plot(fd.sw.lm)
ggplot(flamm.df, aes(x = sample_wt, y = fd)) +
  geom_point() +
  geom_smooth(method = 'lm') +
  labs(x = 'Sample Wt. (g)', y = 'Flame Duration (s)') +
  theme_bw()
## FH ## - SIGNIFICANT
fh.sw.lm <- lm(scale(fh) ~ scale(sample_wt), data = flamm.df)
summary(fh.sw.lm)
plot(fh.sw.lm)
ggplot(flamm.df, aes(x = sample_wt, y = fh)) +
  geom_point() +
  geom_smooth(method = 'lm') +
  labs(x = 'Sample Wt. (g)', y = 'Flame Height (cm)') +
  theme_bw()
## PFG ## - SIGNIFICANT
pfg.sw.lm <- lm(scale(pfg) ~ scale(sample_wt), data = flamm.df)
summary(pfg.sw.lm)
plot(pfg.sw.lm)
ggplot(flamm.df, aes(x = sample_wt, y = pfg)) +
  geom_point() +
  geom_smooth(method = 'lm') +
  labs(x = 'Sample Wt. (g)', y = 'Post-Flame Glow (s)') +
  theme_bw()
## LFM ## - SIGNIFICANT
lfm.sw.lm <- lm(lfm ~ sample_wt, data = flamm.df)
summary(lfm.sw.lm)
plot(lfm.sw.lm)
ggplot(flamm.df, aes(x = sample_wt, y = lfm)) +
  geom_point() +
  geom_smooth(method = 'lm') +
  labs(x = 'Sample Wt. (g)', y = 'Live Fuel Moisture (%)') +
  theme_bw()
## MPa ## - n.s.
mpa.sw.lm <- lm(mpa ~ sample_wt, data = flamm.df)
summary(mpa.sw.lm)
plot(mpa.sw.lm)
ggplot(flamm.df, aes(x = sample_wt, y = mpa)) +
  geom_point() +
  geom_smooth(method = 'lm') +
  labs(x = 'Sample Wt. (g)', y = 'Water Potential (MPa)') +
  theme_bw()
```

## Nonlinear Relationships
Generalized Additive Models 1
```{r}
# FH
flamm.df.temp.fh <- flamm.df %>% 
  drop_na(fh, sample_wt) %>% 
  select(fh, sample_wt) %>% 
  mutate(fh = scale(fh), sample_wt = scale(sample_wt))
fh.sw.gam <- gam(fh ~ s(sample_wt), data = flamm.df.temp.fh)
summary(fh.sw.gam) 
flamm.df.temp.fh$predict.gam.fh <- predict.Gam(fh.sw.gam)
ggplot(data = flamm.df.temp.fh) +
  geom_point(aes(x = sample_wt, y = predict.gam.fh), color = 'black') +
  geom_point(aes(x = sample_wt, y = fh), alpha = 0.5, color = 'gray50')

# FD
flamm.df.temp.fd <- flamm.df %>% 
  drop_na(fd, sample_wt) %>% 
  select(fd, sample_wt) %>% 
  mutate(fd = scale(fd), sample_wt = scale(sample_wt))
fd.sw.gam <- gam(fd ~ s(sample_wt), data = flamm.df.temp.fd)
summary(fd.sw.gam)
flamm.df.temp.fd$predict.gam.fd <- predict.Gam(fd.sw.gam)
ggplot(data = flamm.df.temp.fd) +
  geom_point(aes(x = sample_wt, y = predict.gam.fd), color = 'black') +
  geom_point(aes(x = sample_wt, y = fd), alpha = 0.5, color = 'gray50')

# PFG
flamm.df.temp.pfg <- flamm.df %>% 
  drop_na(pfg, sample_wt) %>% 
  select(pfg, sample_wt) %>% 
  mutate(pfg = scale(pfg), sample_wt = scale(sample_wt))
pfg.sw.gam <- gam(pfg ~ s(sample_wt), data = flamm.df.temp.pfg)
summary(pfg.sw.gam)
flamm.df.temp.pfg$predict.gam.pfg <- predict.Gam(pfg.sw.gam)
ggplot(data = flamm.df.temp.pfg) +
  geom_point(aes(x = sample_wt, y = predict.gam.pfg), color = 'black') +
  geom_point(aes(x = sample_wt, y = pfg), alpha = 0.5, color = 'gray50')
```

Generalized Additive Models 2
```{r}
# FH
fh.sw.gam <- gam(fh ~ lo(sample_wt), data = flamm.df.temp.fh)
summary(fh.sw.gam)
flamm.df.temp.fh$predict.gam.fh <- predict.Gam(fh.sw.gam)
ggplot(data = flamm.df.temp.fh) +
  geom_point(aes(x = sample_wt, y = predict.gam.fh), color = 'black') +
  geom_point(aes(x = sample_wt, y = fh), alpha = 0.5, color = 'gray50')

# FD
fd.sw.gam <- gam(fd ~ lo(sample_wt), data = flamm.df.temp.fd)
summary(fd.sw.gam)
flamm.df.temp.fd$predict.gam.fd <- predict.Gam(fd.sw.gam)
ggplot(data = flamm.df.temp.fd) +
  geom_point(aes(x = sample_wt, y = predict.gam.fd), color = 'black') +
  geom_point(aes(x = sample_wt, y = fd), alpha = 0.5, color = 'gray50')

# PFG
pfg.sw.gam <- gam(pfg ~ lo(sample_wt), data = flamm.df.temp.pfg)
summary(pfg.sw.gam)
flamm.df.temp.pfg$predict.gam.pfg <- predict.Gam(pfg.sw.gam)
ggplot(data = flamm.df.temp.pfg) +
  geom_point(aes(x = sample_wt, y = predict.gam.pfg), color = 'black') +
  geom_point(aes(x = sample_wt, y = pfg), alpha = 0.5, color = 'gray50')
```

Polynomial Regressions
```{r}
# First, testing for which degree gives us best results
fh.poly.aic <- (rep(NA, 5))
for(i in 1:5){
fh.sw.poly <- lm(data = flamm.df.temp.fh, fh ~ poly(sample_wt, i))
fh.poly.aic[i] <- AIC(fh.sw.poly)
}
fh.poly.aic # degrees = 3

fd.poly.aic <- (rep(NA, 5))
for(i in 1:5){
fd.sw.poly <- lm(data = flamm.df.temp.fd, fd ~ poly(sample_wt, i))
fd.poly.aic[i] <- AIC(fd.sw.poly)
}
fd.poly.aic # degrees = 3

pfg.poly.aic <- (rep(NA, 5))
for(i in 1:5){
pfg.sw.poly <- lm(data = flamm.df.temp.pfg, pfg ~ poly(sample_wt, i))
pfg.poly.aic[i] <- AIC(pfg.sw.poly)
}
pfg.poly.aic # linear is best, but since others are degrees = 3, we'll look into that one for all three metrics

# Running models
fh.sw.poly <- lm(data = flamm.df.temp.fh, fh ~ poly(sample_wt, 3))
summary(fh.sw.poly)
fd.sw.poly <- lm(data = flamm.df.temp.fd, fd ~ poly(sample_wt, 3))
summary(fd.sw.poly)
pfg.sw.poly <- lm(data = flamm.df.temp.pfg, pfg ~ poly(sample_wt, 3))
summary(pfg.sw.poly)
```


## Binned Sample Weights
```{r}
# Data wrangling
quantile(flamm.df$sample_wt, probs = c(0, 0.25, 0.5, 0.75, 1), na.rm = TRUE)

flamm.df <- flamm.df %>% 
  mutate(sample_wt_bins = case_when(
    sample_wt <= 1.1064 ~ '0.2264-1.1064',
    sample_wt > 1.1064 & sample_wt <= 2.8064  ~ '1.1064-2.8064',
    sample_wt > 2.8064 & sample_wt <= 8.4864 ~ '2.8064-8.4864',
    sample_wt > 8.4864 & sample_wt <= 30.5664 ~ '8.4864-30.5664',
  ))

# LMA
flamm.df %>% 
  filter(LMA < 2.5) %>% 
  drop_na(sample_wt_bins)%>% 
ggplot(aes(x = LMA, y = fh)) +
  labs(x = 'Leaf Mass per Area', y = 'Flame Height (cm)') +
  geom_point() +
  geom_smooth(method = 'lm', color = 'black') +
  facet_wrap(~sample_wt_bins, scales = 'free_x') +
  theme_bw()

flamm.df %>% 
  filter(LMA < 2.5) %>% 
  drop_na(sample_wt_bins)%>% 
ggplot(aes(x = LMA, y = fd)) +
  labs(x = 'Leaf Mass per Area', y = 'Flame Duration (s)') +
  geom_point() +
  geom_smooth(method = 'lm', color = 'black') +
  facet_wrap(~sample_wt_bins, scales = 'free_x') +
  theme_bw()

# Leaf Mass Ratio
flamm.df %>% 
  drop_na(sample_wt_bins)%>% 
ggplot(aes(x = leaf_mass_ratio, y = fh)) +
  labs(x = 'Leaf Mass Ratio', y = 'Flame Height (cm)') +
  geom_point() +
  geom_smooth(method = 'lm', color = 'black') +
  facet_wrap(~sample_wt_bins, scales = 'free_x') +
  theme_bw()

flamm.df %>% 
  drop_na(sample_wt_bins)%>% 
ggplot(aes(x = leaf_mass_ratio, y = fd)) +
  labs(x = 'Leaf Mass Ratio', y = 'Flame Duration (s)') +
  geom_point() +
  geom_smooth(method = 'lm', color = 'black') +
  facet_wrap(~sample_wt_bins, scales = 'free_x') +
  theme_bw()
```

## Different transformations
```{r}
flamm.df.sw <- flamm.df %>% 
  mutate(fh = scales::rescale(fh),
         fd = scales::rescale(fd),
         pfg = scales::rescale(pfg),
         sample_wt = scales::rescale(sample_wt)) %>% 
  mutate(fh.norm = fh/sample_wt,
         fd.norm = fd/sample_wt,
         pfg.norm = pfg/sample_wt) %>% 
  mutate(fh.norm.sqrt = fh/(1+sample_wt^1/2),
         fd.norm.sqrt = fd/(1+sample_wt^1/2),
         pfg.norm.sqrt = pfg/(1+sample_wt^1/2)) %>% 
  mutate(fh.norm.cube.root = fh/(1+sample_wt^1/3),
         fd.norm.cube.root = fd/(1+sample_wt^1/3),
         pfg.norm.cube.root = pfg/(1+sample_wt^1/3))

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = fh)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = fh.norm)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = fh.norm.sqrt)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = fh.norm.cube.root)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = fd)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = fd.norm)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = fd.norm.sqrt)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = fd.norm.cube.root)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = pfg)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = pfg.norm)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = pfg.norm.sqrt)) +
    geom_point()

flamm.df.sw %>% 
  ggplot(aes(x = sample_wt, y = pfg.norm.cube.root)) +
    geom_point()

flamm.df.sw.long <- flamm.df.sw %>% 
  select(date, species, plant, rep, mpa, leaf_lfm, leaf_lfm, stem_lfm, lfm, LMA, SLA, thickness, sample_wt, fh, fd, pfg, fh.norm, fd.norm, pfg.norm, fh.norm.sqrt, fd.norm.sqrt, pfg.norm.sqrt, fh.norm.cube.root, fd.norm.cube.root, pfg.norm.cube.root) %>% 
  pivot_longer(!c(date, species, plant, rep, mpa, leaf_lfm, leaf_lfm, stem_lfm, lfm, LMA, SLA, thickness, sample_wt), names_to = 'trait', values_to = 'value') %>% 
  mutate(species = factor(species, levels = c('SALAPI', 'SALLEU', 'CEAGRI', 'ARTCAL', 'QUEAGR', 'MALLAU', 'EUCGLO', 'ERIKAR', 'HETARB', 'IRIDOU', 'ARCDEN'))) #note that EUGL needs to change to EUCGLO

flamm.df.sw.long %>% 
ggplot(aes(x = species, y = value, color = species)) +
  geom_boxplot() +
  labs(x = "Species", y = "Trait Value") +
  facet_wrap(~trait, ncol = 4, scales = 'free_y') +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        strip.text = element_text(face = 'bold', size = 16),
        axis.text.x = element_text(angle = 20, size = 12))

flamm.df.sw.long %>% 
ggplot(aes(x = species, y = sample_wt, color = species)) +
  geom_boxplot() +
  labs(x = "Species", y = "Sample Weight (g)") +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        strip.text = element_text(face = 'bold', size = 16),
        axis.text.x = element_text(angle = 20, size = 12))
```

# Sample Dry Weight/Wet Weight
```{r}
# Calculating the metrics
flamm.df <- flamm.df %>% 
  mutate(wet_mass = stem_wet_mass + leaf_wet_mass,
         dry_mass = stem_dry_mass + leaf_dry_mass,
         gdw_gfw = dry_mass/(wet_mass+dry_mass),
         dw_flam_sample = gdw_gfw * sample_wt, # need to change to sample weight
         ww_flam_sample = sample_wt - dw_flam_sample)

# Boxplots
ggplot(flamm.df, aes(x = species, color = species, y = dw_flam_sample)) +
  geom_boxplot(outlier.shape = NA) +
  geom_jitter(width = 0.1, height = 0.01, alpha = 0.4, aes(color = species)) +
  labs(x = "Species", y = "Dry Weight (g)") +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(angle = 20, size = 12))

ggplot(flamm.df, aes(x = species, color = species, y = ww_flam_sample)) +
  geom_boxplot(outlier.shape = NA) +
  geom_jitter(width = 0.1, height = 0.01, alpha = 0.4, aes(color = species)) +
  labs(x = "Species", y = "Wet Weight (g)") +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(angle = 20, size = 12))

# Scatterplots
flamm.df.long.scatter <- flamm.df %>% 
  select(date, species, plant, rep, mpa, branching, branch_height, branch_length, branch_width, leaf_lfm, leaf_lfm, stem_lfm, lfm, LMA, SLA, thickness, sample_wt, fh, fd, tti, pfg, max_temp, max_heat_flux_loessCH7, max_heat_flux_loessCH8, dw_flam_sample, ww_flam_sample) %>%  # took out  gd, gti, ttfg for now
  pivot_longer(!c(date, species, plant, rep, mpa, branching, branch_height, branch_length, branch_width, leaf_lfm, stem_lfm, lfm, LMA, SLA, thickness, sample_wt, dw_flam_sample, ww_flam_sample), names_to = 'trait', values_to = 'value') %>% 
  filter(trait %in% c('fd', 'fh', 'pfg', 'temp_change', 'heat_flux_change')) %>% 
  mutate(species = factor(species, levels = c('SALAPI', 'SALLEU', 'CEAGRI', 'ARTCAL', 'QUEAGR', 'MALLAU', 'EUCGLO', 'ERIKAR', 'HETARB', 'IRIDOU', 'ARCDEN'))) #note that EUGL needs to change to EUCGLO

exploratory_scatterplot(flamm.df.long.scatter$dw_flam_sample, 'Dry Weight (g)')
exploratory_scatterplot(flamm.df.long.scatter$ww_flam_sample, 'Wet Weight (g)')
```

It seems like sample weight is driving these relationships far more than hydration/dryness. Let's try some corrections:
```{r}
ggplot(flamm.df, aes(y = fh/dw_flam_sample, x = dw_flam_sample)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 350)+ 
  stat_regline_equation(label.y = 400) +
  labs(y = "Flame Height/Dry Weight", x = "Dry Weight (g)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

ggplot(flamm.df, aes(y = log(fh/dw_flam_sample), x = dw_flam_sample)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 7)+ 
  stat_regline_equation(label.y = 8) +
  labs(y = "Flame Height/Dry Weight (log-transformed)", x = "Dry Weight (g)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

ggplot(flamm.df, aes(y = fh/1 + dw_flam_sample, x = dw_flam_sample)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 65)+ 
  stat_regline_equation(label.y = 75) +
  labs(y = "Flame Height/1 + Dry Weight", x = "Dry Weight (g)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

flamm.df %>% 
  filter(lfm < 500) %>% 
  ggplot(aes(x = lfm, y = fh/dw_flam_sample)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 175)+ 
  stat_regline_equation(label.y = 200) +
  labs(x = "Live Fuel Moisture (%)", y = "Flame Height/Dry Weight (cm/g)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))
```

By species
```{r}
ggplot(flamm.df, aes(y = fh/dw_flam_sample, x = dw_flam_sample, color = species)) +
  geom_point() +
  geom_smooth(method = 'lm', se = F) +
  stat_cor(label.y = c(450, 428, 406, 384, 363, 362, 340, 318, 296, 274),label.x = 5) + 
  stat_regline_equation(label.y = c(450, 428, 406, 384, 363, 362, 340, 318, 296, 274), 
                        label.x = 1.5) +
  labs(y = "Flame Height/Dry Weight", x = "Dry Weight (g)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

ggplot(flamm.df, aes(y = log(fh/dw_flam_sample), x = dw_flam_sample, color = species)) +
  geom_point() +
  geom_smooth(method = 'lm', se = F) +
  stat_cor(label.y = c(8.5, 8, 7.5, 7, 6.6, 6.5, 6, 5.5, 5, 4.5),label.x = 4.5) + 
  stat_regline_equation(label.y = c(8.5, 8, 7.5, 7, 6.6, 6.5, 6, 5.5, 5, 4.5), label.x = 1.5) +
  labs(y = "Flame Height/Dry Weight (log-transformed)", x = "Dry Weight (g)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

ggplot(flamm.df, aes(y = fh/1 + dw_flam_sample, x = dw_flam_sample, color = species)) +
  geom_point() +
  geom_smooth(method = 'lm', se = F) +
  stat_cor(label.y = c(120, 115, 110, 105, 101, 100, 95, 90, 85, 80),label.x = 3) + 
  stat_regline_equation(label.y = c(120, 115, 110, 105, 101, 100, 95, 90, 85, 80)) +
  labs(y = "Flame Height/1 + Dry Weight", x = "Dry Weight (g)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

flamm.df %>% 
  filter(lfm < 500) %>% 
  ggplot(aes(x = lfm, y = fh/dw_flam_sample, color = species)) +
  geom_point() +
  geom_smooth(method = 'lm', se = F) +
  stat_cor(label.y = c(250, 240, 230, 220, 210, 200, 190, 180, 170, 160),label.x = 200) + 
  stat_regline_equation(label.y = c(250, 240, 230, 220, 210, 200, 190, 180, 170, 160)) +
  labs(x = "Live Fuel Moisture (%)", y = "Flame Height/Dry Weight (cm/g)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))
```

Polynomial Regression - FH vs. DW
```{r}
flamm.df.temp.fh <- flamm.df %>% 
  drop_na(fh, dw_flam_sample)
# First, testing for which degree gives us best results
fh.poly.aic <- (rep(NA, 5))
for(i in 1:5){
fh.sw.poly <- lm(data = flamm.df.temp.fh, fh ~ poly(dw_flam_sample, i))
fh.poly.aic[i] <- AIC(fh.sw.poly)
}
fh.poly.aic # degrees = 5

fh.dw.poly <- lm(data = flamm.df.temp.fh, fh ~ poly(dw_flam_sample, 5))
summary(fh.dw.poly)

flamm.df.temp.fh$predict.poly.fh <- predict(fh.dw.poly)
ggplot(data = flamm.df.temp.fh) +
  geom_point(aes(x = dw_flam_sample, y = predict.poly.fh), color = 'black') +
  geom_point(aes(x = dw_flam_sample, y = fh), alpha = 0.5, color = 'gray50') +
  theme_bw() +
  labs(y = 'Flame Height (cm)', x = 'Dry Weight (g)') +
  annotate('text', x = 4, y = 80, label = 'Note: black points = polynomial regression') +
  annotate('text', x = 4, y = 75, label = 'Adjusted R^2 = 0.537, p < 2.2e-16') +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

# Residuals
flamm.df.temp.fh <- flamm.df.temp.fh %>% 
  mutate(fh.dev.poly = fh - predict.poly.fh)

flamm.df.temp.fh %>% 
  ggplot(aes(x = dw_flam_sample, y = fh.dev.poly)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 27)+ 
  stat_regline_equation(label.y = 30) +
  labs(x = "Dry Weight (g)", y = "FH~DW Polynomial Residuals") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

flamm.df.temp.fh %>% 
  ggplot(aes(x = mpa, y = fh.dev.poly)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 27)+ 
  stat_regline_equation(label.y = 30) +
  labs(x = "Water Potential (MPa)", y = "FH~DW Polynomial Residuals") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

flamm.df.temp.fh %>% 
  ggplot(aes(x = mpa, y = fh)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 75)+ 
  stat_regline_equation(label.y = 80) +
  labs(x = "Water Potential (MPa)", y = "Flame Height (cm)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

flamm.df.temp.fh %>% 
  mutate(species = factor(species, levels = c('IRIDOU', 'CEAGRI', 'ARTCAL', 'ARCDEN', 'SALAPI', 'MALLAU', 'SALLEU', 'ERIKAR', 'HETARB', 'QUEAGR'))) %>% 
  ggplot(aes(x = species, y = fh.dev.poly)) + 
    geom_jitter(width = 0.2, aes(color = species)) +
    geom_boxplot(alpha = 0.2, outlier.shape = NA) +
    labs(x = ' ', y = 'FH~DW Polynomial Residuals') +
    theme_bw() +
    theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12, angle = 15))
```


# Plant Traits Exploratory Analyses
Zoe is working on leaf area, LMA, SLA, etc. and I'm focusing on branching, branch width, length, VOCs
## Branching
```{r}
exploratory_scatterplot(flamm.df.long.scatter$branching, 'Branching') # right away, it looks like maybe flame height could relate to branching

ggplot(flamm.df, aes(x = branching, y = fh)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  scale_x_continuous(limits = c(0, 16)) + # Changing limits of x axis
  stat_cor(label.y = 72)+  # To display r value and p value
  stat_regline_equation(label.y = 80) + # To display equation
  labs(x = "Branching", y = "Flame Height (cm)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16), # Bolding axis titles
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

ggplot(flamm.df, aes(x = branching, y = fd)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  scale_x_continuous(limits = c(0, 16)) +
  stat_cor(label.y = 90)+ 
  stat_regline_equation(label.y = 95) +
  labs(x = "Branching", y = "Flame Duration (s)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

ggplot(flamm.df, aes(x = species, y = branching, color = species)) +
  geom_boxplot(outlier.shape = NA) + #
  geom_jitter(width = 0.1, height = 0.01, alpha = 0.4, aes(color = species)) + # To add points behind boxplots
  labs(x = " ", y = "Sample Volume") +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12, angle = 20))

flamm.df %>% 
  filter(species != 'IRIDOU') %>% 
  ggplot(aes(x = branching, y = fh)) +
    geom_point() +
    geom_smooth(method = 'lm', color = 'gray40') +
    facet_wrap(~species, scales = 'free_x') +
    stat_cor(label.y = 72)+ 
    stat_regline_equation(label.y = 80) +
    labs(x = "Branching", y = "Flame Height (cm)") +
    theme_bw() +
    theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))
```

## Sample Volume
```{r}
flamm.df <- flamm.df %>% 
  mutate(branch_volume = branch_length*branch_width*branch_height)

ggplot(flamm.df, aes(x = species, y = branch_volume, color = species)) +
  geom_boxplot(outlier.shape = NA) +
  geom_jitter(width = 0.1, height = 0.01, alpha = 0.4, aes(color = species)) +
  labs(x = " ", y = "Sample Volume") +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12, angle = 20))

ggplot(flamm.df, aes(x = species, y = log(branch_volume), color = species)) +
  geom_boxplot(outlier.shape = NA) +
  geom_jitter(width = 0.1, height = 0.01, alpha = 0.4, aes(color = species)) +
  labs(x = " ", y = "Sample Volume (log-transformed)") +
  theme_bw() +
  theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12, angle = 20))

ggplot(flamm.df, aes(x = branch_volume, y = fh)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 72)+ 
  stat_regline_equation(label.y = 80) +
  labs(x = "Sample Volume", y = "Flame Height (cm)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

ggplot(flamm.df, aes(x = branch_volume, y = fd)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 90)+ 
  stat_regline_equation(label.y = 95) +
  labs(x = "Sample Volume", y = "Flame Duration (s)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))
```

## Compare to LFM, MPa
```{r}
ggplot(flamm.df, aes(x = lfm, y = fh)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 72)+ 
  stat_regline_equation(label.y = 80) +
  labs(x = "Live Fuel Moisture", y = "Flame Height (cm)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

ggplot(flamm.df, aes(x = lfm, y = fd)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 90)+ 
  stat_regline_equation(label.y = 95) +
  labs(x = "Live Fuel Moisture", y = "Flame Duration (s)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

ggplot(flamm.df, aes(x = mpa, y = fh)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 72)+ 
  stat_regline_equation(label.y = 80) +
  labs(x = "Water Potential", y = "Flame Height (cm)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

ggplot(flamm.df, aes(x = mpa, y = fd)) +
  geom_point() +
  geom_smooth(method = 'lm', color = 'gray40') +
  stat_cor(label.y = 90)+ 
  stat_regline_equation(label.y = 95) +
  labs(x = "Water Potential", y = "Flame Duration (s)") +
  theme_bw() +
  theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))
```

## Dry Matter Content
```{r}

```


## Correlations
```{r}
flamm.df %>% 
  filter(lfm < 500) %>% 
  filter(temp_change < 6000 & temp_change > -6000) %>% # getting rid of NA, -Inf, Inf values
  select(fh, fd, pfg, temp_change, heat_flux_change, prop_ig, lfm, leaf_lfm, stem_lfm, LMA, thickness, branch_height, branch_width, branch_length, branch_volume, branching, mpa, sample_wt, stem_mass_ratio, thickness, auc_voc, avg_voc) %>% 
  ggcorr(label = T)
```

Looking into strange correlations/lack of correlations
```{r}
flamm.df %>% 
  filter(lfm < 500) %>% 
    ggplot(aes(x = thickness, y = prop_ig)) +
      geom_point(alpha = 0.4) +
      geom_smooth()

flamm.df %>% 
  filter(start_temp > 10) %>% 
ggplot(aes(x = date, y = start_temp)) +
  geom_jitter(width = 0.1, height = 0.1, aes(color = species)) +
  geom_violin(alpha = 0.4) +
  theme_bw() +
  labs(x = 'Date', y = 'Starting Temp. (C)', color = 'Species')

flamm.df %>% 
ggplot(aes(x = as.factor(pre_ignition_glow), y = fd)) +
  facet_wrap(~species, scales = 'free_y') +
  geom_jitter(width = 0.1, height = 0.1, aes(color = species)) +
  geom_boxplot(alpha = 0.4) +
  theme_bw() +
  labs(x = 'Pre-ignition Glow?', y = 'Flame Duration (s)', color = 'Species')
```

# VOCs
```{r}
flamm.df %>% 
  pivot_longer(cols = c(fh, fd, pfg, prop_ig, temp_change, heat_flux_change), names_to = 'var', values_to = 'val') %>% 
  ggplot(aes(x = auc_voc, y = val, color = species)) +
    geom_point() +
    geom_smooth(method = 'lm', color = 'gray30', alpha = 0.8) +
    facet_wrap(~var, scales = 'free', labeller = labeller(var = flam_labels)) + 
    labs(x = 'VOCs (Area Under Curve)', y = 'Flammability Metric Value', color = 'Species') +
    theme_bw() +
    theme(legend.position = 'right',
          legend.title = element_text(face = 'bold', size = 15),
          axis.title = element_text(face = 'bold', size = 17),
          axis.text.y = element_text(size = 15),
          axis.text.x = element_text(angle = 15, size = 12),
          strip.text = element_text(face = 'bold', size = 15))

flamm.df %>% 
  pivot_longer(cols = c(fh, fd, pfg, prop_ig, temp_change, heat_flux_change), names_to = 'var', values_to = 'val') %>% 
  ggplot(aes(x = log(avg_voc), y = val, color = species)) +
    geom_point() +
    geom_smooth(method = 'lm', color = 'gray30', alpha = 0.8) +
    facet_wrap(~var, scales = 'free', labeller = labeller(var = flam_labels)) + 
    labs(x = 'VOCs/second (log-transformed)', y = 'Flammability Metric Value', color = 'Species') +
    theme_bw() +
    theme(legend.position = 'right',
          legend.title = element_text(face = 'bold', size = 15),
          axis.title = element_text(face = 'bold', size = 17),
          axis.text.y = element_text(size = 15),
          axis.text.x = element_text(size = 12),
          strip.text = element_text(face = 'bold', size = 15))

flamm.df %>% 
  pivot_longer(cols = c(fh, fd, pfg, prop_ig, temp_change, heat_flux_change), names_to = 'var', values_to = 'val') %>% 
  ggplot(aes(x = log(norm_auc_voc), y = val, color = species)) +
    geom_point() +
    geom_smooth(method = 'lm', color = 'gray30', alpha = 0.8) +
    facet_wrap(~var, scales = 'free', labeller = labeller(var = flam_labels)) + 
    labs(x = 'VOCs (AUC; Normalized, log-transformed)', y = 'Flammability Metric Value', color = 'Species') +
    theme_bw() +
    theme(legend.position = 'right',
          legend.title = element_text(face = 'bold', size = 15),
          axis.title = element_text(face = 'bold', size = 17),
          axis.text.y = element_text(size = 15),
          axis.text.x = element_text(angle = 15, size = 12),
          strip.text = element_text(face = 'bold', size = 15))

flamm.df %>% 
  pivot_longer(cols = c(fh, fd, pfg, prop_ig, temp_change, heat_flux_change), names_to = 'var', values_to = 'val') %>% 
  ggplot(aes(x = log(norm_avg_voc), y = val, color = species)) +
    geom_point() +
    geom_smooth(method = 'lm', color = 'gray30', alpha = 0.8) +
    facet_wrap(~var, scales = 'free', labeller = labeller(var = flam_labels)) + 
    labs(x = 'VOCs/second (Normalized, log-transformed)', y = 'Flammability Metric Value', color = 'Species') +
    theme_bw() +
    theme(legend.position = 'right',
          legend.title = element_text(face = 'bold', size = 15),
          axis.title = element_text(face = 'bold', size = 17),
          axis.text.y = element_text(size = 15),
          axis.text.x = element_text(size = 12),
          strip.text = element_text(face = 'bold', size = 15))
```


# Sample Size vs. FD/FH
```{r}
size_labels <- c('Sample Weight (g)', 'Dry Weight (g)', 'Wet Weight (g)', 'Sample Volume (cm3)')
names(size_labels) <- c('sample_wt', 'dw_flam_sample', 'ww_flam_sample', 'branch_volume')

# All species
flamm.df %>% 
  select(species, fh, fd, sample_wt, dw_flam_sample, ww_flam_sample, branch_volume) %>% 
  pivot_longer(cols = !c(fh, fd, species), names_to = 'size_var', values_to = 'size_value') %>%
  pivot_longer(cols = c(fh, fd), names_to = 'flam_var', values_to = 'flam_value') %>% 
  na.omit() %>% 
  ggplot(aes(x = size_value, y = flam_value, color = species)) +
    geom_point(alpha = 0.5) +
    geom_smooth(alpha = 0.6, linewidth = 0.5, se = F, method = 'lm') +
    facet_grid(flam_var ~ size_var, scales = 'free', labeller = labeller(flam_var = flam_labels, size_var = size_labels)) +
    labs(x = 'Sample Size Metric', y = ' ', color = 'Species') +
    theme_bw() +
    theme(axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        legend.title = element_text(face = 'bold', size = 16),
        legend.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))

# By species
spp_vector <- c('SALAPI', 'SALLEU', 'CEAGRI', 'ARTCAL', 'MALLAU', 'ERIKAR')
plot_list <- list()
for(i in 1:length(spp_vector)) {
plot <- flamm.df %>% 
  filter(species == spp_vector[i]) %>% 
  select(fh, fd, sample_wt, dw_flam_sample, ww_flam_sample, branch_volume) %>% 
    pivot_longer(cols = !c(fh, fd), names_to = 'size_var', values_to = 'size_value') %>%
  pivot_longer(cols = c(fh, fd), names_to = 'flam_var', values_to = 'flam_value') %>% 
  na.omit() %>% 
  ggplot(aes(x = size_value, y = flam_value)) +
    geom_point(alpha = 0.5) +
    geom_smooth(alpha = 0.6, linewidth = 0.5, method = 'lm', color = 'gray30') +
    facet_grid(flam_var ~ size_var, scales = 'free', labeller = labeller(flam_var = flam_labels, size_var = size_labels)) +
    labs(x = 'Sample Size Metric', y = ' ', title = paste0(spp_vector[i], ': Sample Size vs. Flammability Metrics')) +
    theme_bw() +
    theme(legend.position = 'none',
        axis.title = element_text(face = 'bold', size = 16),
        axis.text = element_text(size = 14),
        axis.text.x = element_text(size = 12))
plot_list[[i]] <- plot
}
plot_list[[1]]
plot_list[[2]]
plot_list[[3]]
plot_list[[4]]
plot_list[[5]]
plot_list[[6]]
```
