==> scripts/2.2_sample_weight_exploration.Rmd <==
---
title: "Sample Weight Exploration"
author: "Joe Celebrezze"
date: "2023-10-09"
output: html_document
---

# Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(here)
here = here::here
source(here("scripts", "source_code.R"))
source(here('scripts', 'mem.selection.function.R')) #this is where mem.selection(), mem.selection.table() and mallows.cp() functions are stored

# only positive leaf lfm values
flamm.df <- flamm.df %>% 
  filter(leaf_lfm > 0)
```

# Scatterplots
```{r}
flamm.df %>% 
  pivot_longer(cols = c(leaf_mass_ratio, leaf_lfm, stem_lfm, lfm, leaf_dmc, stem_dmc, dmc, LMA, branch_volume, branching, mpa, no_branches), names_to = 'exp_var', values_to = 'exp_val') %>% 
  mutate(exp_var = fct_relevel(as.factor(exp_var), c('leaf_mass_ratio', 'branch_volume', 'branching', 'no_branches', 'leaf_lfm', 'stem_lfm', 'lfm', 'mpa', 'leaf_dmc', 'stem_dmc', 'dmc', 'LMA'))) %>% 
  ggplot(aes(y = sample_wt, x = exp_val, color = species)) +
    geom_point() +
    geom_smooth(method = 'lm', se = F) +
    facet_wrap(~exp_var, scales = 'free') +
    scale_color_manual(values = c('#00468BB2', '#ED0000B2', '#42B540B2', '#0099B4B2', '#925E9FB2', '#AD002AB2', '#ADB6B6B2', '#1B1919B2')) +
    theme_bw() +
    labs(y = 'Sample Weight (g)', x = 'Explanatory Variable', color = 'Species') +
    theme(legend.position = 'top',
          axis.title = element_text(face = 'bold', size = 16),
          axis.text = element_text(size = 14),
          legend.title = element_text(face = 'bold', size = 16),
          legend.text = element_text(size = 14))
ggsave(here('figures', 'EDA', 'sample_wt', 'regressions.png'), height = 10, width = 12)
```

## By Species
```{r}
explanatory_var <- c('leaf_mass_ratio', 'branch_volume', 'branching', 'no_branches', 'leaf_lfm', 'stem_lfm', 'lfm', 'mpa', 'leaf_dmc', 'stem_dmc', 'dmc', 'LMA')
explanatory_labs <- c('Leaf Mass Ratio', 'Branch Volume', 'Branching', '# of Branches', 'Leaf LFM', 'Stem LFM', 'LFM', 'Water Potential', 'Leaf DMC', 'Stem DMC', 'DMC', 'Leaf Mass per Area')

sw.df <- flamm.df %>% 
  select(leaf_mass_ratio, branch_volume, branching, no_branches, leaf_lfm, stem_lfm, lfm, mpa, leaf_dmc, stem_dmc, dmc, LMA, species, sample_wt)

for(i in 1:length(explanatory_var)){
sw.df %>% 
  ggplot(aes(y = sample_wt, x = sw.df[,i], color = species)) +
    geom_point() +
    geom_smooth(method = 'lm') +
    facet_wrap(~species, scales = 'free') +
    scale_color_manual(values = c('#00468BB2', '#ED0000B2', '#42B540B2', '#0099B4B2', '#925E9FB2', '#AD002AB2', '#ADB6B6B2', '#1B1919B2')) +
    theme_bw() +
    labs(y = 'Sample Weight (g)', x = explanatory_labs[i], color = 'Species') +
    theme(legend.position = 'top',
          axis.title = element_text(face = 'bold', size = 16),
          axis.text = element_text(size = 14),
          legend.title = element_text(face = 'bold', size = 16),
          legend.text = element_text(size = 14))
ggsave(here('figures', 'EDA', 'sample_wt', paste0('spp_', explanatory_var[i],'.png')), height = 10, width = 12)
}
```

## Wet Weight and Dry Weight
```{r}
flamm.df %>% 
  mutate(ww_flam_sample = scale(ww_flam_sample), dw_flam_sample = scale(dw_flam_sample),
         sample_wt = scale(sample_wt)) %>% 
  pivot_longer(cols = c(ww_flam_sample, dw_flam_sample, sample_wt), names_to = 'sw_var', values_to = 'sw_val') %>% 
  pivot_longer(cols = c(fh, fd, temp_change, heat_flux_change), names_to = 'flam_var', values_to = 'flam_val') %>% 
  mutate(flam_var = case_when(flam_var == 'fh' ~ 'Flame Height',
                              flam_var == 'fd' ~ 'Flame Duration',
                              flam_var == 'temp_change' ~ 'Temp. Change',
                              flam_var == 'heat_flux_change' ~ 'Heat Flux Change')) %>% 
  mutate(sw_var = case_when(sw_var == 'ww_flam_sample' ~ 'Wet Weight',
                            sw_var == 'dw_flam_sample' ~ 'Dry Weight',
                            sw_var == 'sample_wt' ~ 'Sample Weight')) %>% 
  ggplot(aes(y = flam_val, x = sw_val, color = sw_var)) +
    geom_point() +
    geom_smooth(method = 'lm') +
    facet_wrap(~flam_var, scales = 'free') +
    scale_color_manual(values = c('#00468BB2', '#ED0000B2', '#42B540B2')) +
    theme_bw() +
    labs(x = 'Weight Components', y = 'Flammability Metrics', color = 'Species') +
    theme(legend.position = 'top',
          axis.title = element_text(face = 'bold', size = 16),
          axis.text = element_text(size = 14),
          legend.title = element_text(face = 'bold', size = 16),
          legend.text = element_text(size = 14))
```

## Leaf Weight and Stem Weight
```{r}
flamm.df <- flamm.df %>% 
  mutate(leaf_wt = sample_wt*leaf_mass_ratio, stem_wt = sample_wt*stem_mass_ratio)

flamm.df %>% 
  mutate(leaf_wt = scale(leaf_wt), stem_wt = scale(stem_wt),
         sample_wt = scale(sample_wt)) %>% 
  pivot_longer(cols = c(leaf_wt, stem_wt, sample_wt), names_to = 'sw_var', values_to = 'sw_val') %>% 
  pivot_longer(cols = c(fh, fd, temp_change, heat_flux_change), names_to = 'flam_var', values_to = 'flam_val') %>% 
  mutate(flam_var = case_when(flam_var == 'fh' ~ 'Flame Height',
                              flam_var == 'fd' ~ 'Flame Duration',
                              flam_var == 'temp_change' ~ 'Temp. Change',
                              flam_var == 'heat_flux_change' ~ 'Heat Flux Change')) %>% 
  mutate(sw_var = case_when(sw_var == 'leaf_wt' ~ 'Leaf Weight',
                            sw_var == 'stem_wt' ~ 'Stem Weight',
                            sw_var == 'sample_wt' ~ 'Sample Weight')) %>% 
  ggplot(aes(y = flam_val, x = sw_val, color = sw_var)) +
    geom_point() +
    geom_smooth(method = 'lm') +
    facet_wrap(~flam_var, scales = 'free') +
    scale_color_manual(values = c('#00468BB2', '#ED0000B2', '#42B540B2')) +
    theme_bw() +
    labs(x = 'Weight Components', y = 'Flammability Metrics', color = 'Species') +
    theme(legend.position = 'top',
          axis.title = element_text(face = 'bold', size = 16),
          axis.text = element_text(size = 14),
          legend.title = element_text(face = 'bold', size = 16),
          legend.text = element_text(size = 14))
```

## Leaf and Stem, Wet and Dry Weight
```{r}
flamm.df <- flamm.df %>% 
  mutate(leaf_ww = ww_flam_sample*leaf_mass_ratio, stem_ww = ww_flam_sample*stem_mass_ratio,
         leaf_dw = dw_flam_sample*leaf_mass_ratio, stem_dw = dw_flam_sample*stem_mass_ratio,) %>% 
  mutate(leaf_ww = scale(leaf_ww), leaf_dw = scale(leaf_dw),
         stem_ww = scale(stem_ww), stem_dw = scale(stem_dw),
         sample_wt = scale(sample_wt))

flamm.df %>% 
  pivot_longer(cols = c(leaf_ww, stem_ww, leaf_dw, stem_dw, sample_wt), names_to = 'sw_var', values_to = 'sw_val') %>% 
  pivot_longer(cols = c(fh, fd, temp_change, heat_flux_change), names_to = 'flam_var', values_to = 'flam_val') %>% 
  mutate(flam_var = case_when(flam_var == 'fh' ~ 'Flame Height',
                              flam_var == 'fd' ~ 'Flame Duration',
                              flam_var == 'temp_change' ~ 'Temp. Change',
                              flam_var == 'heat_flux_change' ~ 'Heat Flux Change')) %>% 
  mutate(sw_var = case_when(sw_var == 'leaf_ww' ~ 'Leaf Wet Weight',
                            sw_var == 'stem_ww' ~ 'Stem Wet Weight',
                            sw_var == 'leaf_dw' ~ 'Leaf Dry Weight',
                            sw_var == 'stem_dw' ~ 'Stem Dry Weight',
                            sw_var == 'sample_wt' ~ 'Sample Weight')) %>% 
  ggplot(aes(y = flam_val, x = sw_val, color = sw_var)) +
    geom_point() +
    geom_smooth(method = 'lm') +
    facet_wrap(~flam_var, scales = 'free') +
    scale_color_manual(values = c("#00468BB2", "#1B1919B2", "#42B540B2", "#0099B4B2", "#ED0000B2")) +
    theme_bw() +
    labs(x = 'Weight Components', y = 'Flammability Metrics', color = 'Species') +
    theme(legend.position = 'top',
          axis.title = element_text(face = 'bold', size = 16),
          axis.text = element_text(size = 14),
          legend.title = element_text(face = 'bold', size = 16),
          legend.text = element_text(size = 14))
```

# MEM
## Data Wrangling
```{r}
mem.df <- flamm.df %>% 
  filter(ignition != '0') %>% 
  filter(species != 'ARCDEN', species != 'HETARB') %>% 
  unite('plant_id', c(species, plant), sep = '_', remove = F) %>% 
  select(species, fh, fd, temp_change, heat_flux_change, sample_wt, leaf_mass_ratio, leaf_lfm, stem_lfm, lfm, leaf_dmc, stem_dmc, dmc, LMA, branch_volume, branching, mpa, no_branches, ww_flam_sample, dw_flam_sample, leaf_wt, stem_wt, leaf_ww, stem_ww, leaf_dw, stem_dw, plant_id) %>% 
  mutate(fh = scale(fh), fd = scale(fd), temp_change = scale(temp_change), heat_flux_change = scale(heat_flux_change), sample_wt = scale(sample_wt), leaf_mass_ratio = scale(leaf_mass_ratio), leaf_lfm = scale(leaf_lfm), stem_lfm = scale(stem_lfm), lfm = scale(lfm), leaf_dmc = scale(leaf_dmc), stem_dmc = scale(stem_dmc), dmc = scale(dmc), LMA = scale(LMA), branch_volume = scale(branch_volume), branching = scale(branching), mpa = scale(mpa), ww_flam_sample = scale(ww_flam_sample), dw_flam_sample = scale(dw_flam_sample), no_branches = scale(no_branches), leaf_wt = scale(leaf_wt), stem_wt = scale(stem_wt), leaf_ww = scale(leaf_ww), stem_ww = scale(stem_ww), leaf_dw = scale(leaf_dw), stem_dw = scale(stem_dw)) %>% 
  drop_na(leaf_mass_ratio, leaf_lfm, stem_lfm, lfm, leaf_dmc, stem_dmc, dmc, LMA, branch_volume, branching, mpa, no_branches, leaf_wt, stem_wt, leaf_ww, stem_ww, leaf_dw, stem_dw)

mem.predictors.df <- mem.df %>% 
  select(leaf_mass_ratio, leaf_lfm, stem_lfm, lfm, leaf_dmc, stem_dmc, dmc, LMA, branch_volume, branching, mpa, no_branches, species)

mem2.predictors.df <- mem.df %>% 
  select(leaf_mass_ratio, leaf_lfm, stem_lfm, leaf_dmc, stem_dmc, species)
```

## Model Selection 1
Sample Weight vs. predictors
```{r}
sw.mem.selection <- mem.selection('sample_wt', mem.predictors.df, mem.df)
sw.model.df <- sw.mem.selection[[1]]
sw.mod.list <- sw.mem.selection[[2]]
sw.top.mem <- mem.selection.table(sw.model.df, sw.mod.list, 'SampleWt_MEM.html') %>% 
  mutate(flam.var = 'sample_wt')
```

## Model Selection 2
Sample weight vs. predictors (no branching or branch volume)
```{r}
sw2.mem.selection <- mem.selection('sample_wt', mem2.predictors.df, mem.df)
sw2.model.df <- sw2.mem.selection[[1]]
sw2.mod.list <- sw2.mem.selection[[2]]
sw2.top.mem <- mem.selection.table(sw2.model.df, sw2.mod.list, 'SampleWt_MEM.html') %>% 
  mutate(flam.var = 'sample_wt')
```

## Model Selection 3
Flammability metrics vs. ww and dw
```{r}
mem.df <- mem.df %>% 
  drop_na(fd, fh, temp_change, heat_flux_change)

fh.mod <- lmer(fh ~ ww_flam_sample + dw_flam_sample + (1|plant_id), mem.df)
fd.mod <- lmer(fd ~ ww_flam_sample + dw_flam_sample + (1|plant_id), mem.df)
temp.mod <- lmer(temp_change ~ ww_flam_sample + dw_flam_sample + (1|plant_id), mem.df)
hf.mod <- lmer(heat_flux_change ~ ww_flam_sample + dw_flam_sample + (1|plant_id), mem.df)

# abridged
tab_model(fh.mod, fd.mod, temp.mod, hf.mod,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          string.pred = "Coeffcient", 
         title = "MEM Selection: Wet Weight and Dry Weight vs. Flam.",
  string.p = "P-Value", 
  p.style = "stars")

fh.mod2 <- lmer(fh ~ ww_flam_sample + (1|plant_id), mem.df)
fd.mod2 <- lmer(fd ~ ww_flam_sample + (1|plant_id), mem.df)
temp.mod2 <- lmer(temp_change ~ ww_flam_sample + (1|plant_id), mem.df)
hf.mod2 <- lmer(heat_flux_change ~ ww_flam_sample  + (1|plant_id), mem.df)
fh.mod3 <- lmer(fh ~ dw_flam_sample + (1|plant_id), mem.df)
fd.mod3 <- lmer(fd ~ dw_flam_sample + (1|plant_id), mem.df)
temp.mod3 <- lmer(temp_change ~ dw_flam_sample + (1|plant_id), mem.df)
hf.mod3 <- lmer(heat_flux_change ~ dw_flam_sample + (1|plant_id), mem.df)

# elongated
tab_model(fh.mod, fh.mod2, fh.mod3, fd.mod, fd.mod2, fd.mod3,
          temp.mod, temp.mod2, temp.mod3, hf.mod, hf.mod2, hf.mod3,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          string.pred = "Coeffcient", 
         title = "MEM Selection: Wet Weight and Dry Weight vs. Flam.",
  string.p = "P-Value", 
  p.style = "stars")
```

## Model Selection 4
Flammability metrics vs. leaf wt and stem wt
```{r}
mem.df <- mem.df %>% 
  drop_na(fd, fh, temp_change, heat_flux_change)

fh.mod <- lmer(fh ~ leaf_wt + stem_wt + (1|plant_id), mem.df)
fd.mod <- lmer(fd ~ leaf_wt + stem_wt + (1|plant_id), mem.df)
temp.mod <- lmer(temp_change ~ leaf_wt + stem_wt + (1|plant_id), mem.df)
hf.mod <- lmer(heat_flux_change ~ leaf_wt + stem_wt + (1|plant_id), mem.df)

# abridged
tab_model(fh.mod, fd.mod, temp.mod, hf.mod,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          string.pred = "Coeffcient", 
         title = "MEM Selection: Wet Weight and Dry Weight vs. Flam.",
  string.p = "P-Value", 
  p.style = "stars")

fh.mod2 <- lmer(fh ~ leaf_wt + (1|plant_id), mem.df)
fd.mod2 <- lmer(fd ~ leaf_wt + (1|plant_id), mem.df)
temp.mod2 <- lmer(temp_change ~ leaf_wt + (1|plant_id), mem.df)
hf.mod2 <- lmer(heat_flux_change ~ leaf_wt  + (1|plant_id), mem.df)
fh.mod3 <- lmer(fh ~ stem_wt + (1|plant_id), mem.df)
fd.mod3 <- lmer(fd ~ stem_wt + (1|plant_id), mem.df)
temp.mod3 <- lmer(temp_change ~ stem_wt + (1|plant_id), mem.df)
hf.mod3 <- lmer(heat_flux_change ~ stem_wt + (1|plant_id), mem.df)

# elongated
tab_model(fh.mod, fh.mod2, fh.mod3, fd.mod, fd.mod2, fd.mod3,
          temp.mod, temp.mod2, temp.mod3, hf.mod, hf.mod2, hf.mod3,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          string.pred = "Coeffcient", 
         title = "MEM Selection: Leaf Weight and Stem Weight vs. Flam.",
  string.p = "P-Value", 
  p.style = "stars")
```

## Model Selection 5
Flammability metrics vs. leaf wet wt, dry wt and stem wet wt, dry wt
```{r}
mem.df <- mem.df %>% 
  drop_na(fd, fh, temp_change, heat_flux_change)

fh.mod <- lmer(fh ~ leaf_ww + stem_ww + leaf_dw + stem_dw + (1|plant_id), mem.df)
fd.mod <- lmer(fd ~ leaf_ww + stem_ww + leaf_dw + stem_dw + (1|plant_id), mem.df)
temp.mod <- lmer(temp_change ~ leaf_ww + stem_ww + leaf_dw + stem_dw + (1|plant_id), mem.df)
hf.mod <- lmer(heat_flux_change ~ leaf_ww + stem_ww + leaf_dw + stem_dw + (1|plant_id), mem.df)

# abridged
tab_model(fh.mod, fd.mod, temp.mod, hf.mod,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          string.pred = "Coeffcient", 
         title = "MEM Selection: Wet Weight and Dry Weight vs. Flam.",
  string.p = "P-Value", 
  p.style = "stars")

fh.mod2 <- lmer(fh ~ leaf_ww + (1|plant_id), mem.df)
fd.mod2 <- lmer(fd ~ leaf_ww + (1|plant_id), mem.df)
temp.mod2 <- lmer(temp_change ~ leaf_ww + (1|plant_id), mem.df)
hf.mod2 <- lmer(heat_flux_change ~ leaf_ww  + (1|plant_id), mem.df)
fh.mod3 <- lmer(fh ~ stem_ww + (1|plant_id), mem.df)
fd.mod3 <- lmer(fd ~ stem_ww + (1|plant_id), mem.df)
temp.mod3 <- lmer(temp_change ~ stem_ww + (1|plant_id), mem.df)
hf.mod3 <- lmer(heat_flux_change ~ stem_ww + (1|plant_id), mem.df)
fh.mod4 <- lmer(fh ~ leaf_dw + (1|plant_id), mem.df)
fd.mod4 <- lmer(fd ~ leaf_dw + (1|plant_id), mem.df)
temp.mod4 <- lmer(temp_change ~ leaf_dw + (1|plant_id), mem.df)
hf.mod4 <- lmer(heat_flux_change ~ leaf_dw  + (1|plant_id), mem.df)
fh.mod5 <- lmer(fh ~ stem_dw + (1|plant_id), mem.df)
fd.mod5 <- lmer(fd ~ stem_dw + (1|plant_id), mem.df)
temp.mod5 <- lmer(temp_change ~ stem_dw + (1|plant_id), mem.df)
hf.mod5 <- lmer(heat_flux_change ~ stem_dw + (1|plant_id), mem.df)

# elongated
tab_model(fh.mod, fh.mod2, fh.mod3, fh.mod4, fh.mod5,
          fd.mod, fd.mod2, fd.mod3, fd.mod4, fd.mod5,
          temp.mod, temp.mod2, temp.mod3, temp.mod4, temp.mod5,
          hf.mod, hf.mod2, hf.mod3, hf.mod4, hf.mod5,
          show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
          string.pred = "Coeffcient", 
         title = "MEM Selection: Leaf Weight (Wet and Dry) and Stem Weight (Wet and Dry) vs. Flam.",
  string.p = "P-Value", 
  p.style = "stars")
```
