==> scripts/6_conceptual_figures.Rmd <==
---
title: "Conceptual Figures"
author: "Joe Celebrezze"
date: "2023-09-15"
output: html_document
---
# Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(here)
here = here::here
source(here("scripts", "source_code.R"))
```

# Fire Triangles
For these fire triangles, top-performing models were used and fixed and random effects were removed from flam. metrics w/ 'remef' prior to calculating ignitability, combustibility, and consumability scores
*These are the fire triangles we preferred; we also visualized fire triangles w/o 'remef', and w/ 'remef' removing the fixed effects and random effects from 'maximal models'. The same holds true for the Flammability vs. Desirable Traits figures*
```{r}
interspecific_flam_remef <- read.csv(here('data', 'processed-data', 'interspecific_flam_remef.csv'))

r.flam.index.top <- interspecific_flam_remef %>% 
  drop_na(r_fh, r_fd, prop_ig, r_temp_change, r_heat_flux_change) %>% 
  group_by(species) %>% 
  summarise(fh = mean(r_fh), fd = mean(r_fd), prop_ig = mean(prop_ig),
            temp_change = mean(r_temp_change), heat_flux_change = mean(r_heat_flux_change)) %>% 
  mutate(fh = scale(fh), fd = scale(fd), prop_ig = scale(prop_ig),
            temp_change = scale(temp_change), heat_flux_change = scale(heat_flux_change))

min_fh <- min(r.flam.index.top$fh)
min_fd <- min(r.flam.index.top$fd)
min_pi <- min(r.flam.index.top$prop_ig)
min_temp <- min(r.flam.index.top$temp_change)
min_hf <- min(r.flam.index.top$heat_flux_change)

r.flamm.index.top <- r.flam.index.top %>% # so that minimum possible value is 0, subtracting minimum from each flam. metric
  mutate(fh = fh - min_fh,
         fd = fd - min_fd,
         prop_ig = prop_ig - min_pi,
         temp_change = temp_change - min_temp,
         heat_flux_change = heat_flux_change - min_hf)%>% 
  mutate(ignitability = prop_ig, combustibility = (fh + heat_flux_change + temp_change)/3, sustainability = fd) %>% 
  dplyr::select(species, ignitability, combustibility, sustainability) %>% 
  ungroup() %>% 
  mutate(flam.score = (ignitability + combustibility + sustainability)/ 3) %>% 
  mutate(species = case_when(species == 'ARCDEN' ~ 'A. densiflora',
                             species == 'CEAGRI' ~ 'C. griseus',
                             species == 'ARTCAL' ~ 'A. californica',
                             species == 'ERIKAR' ~ 'E. karvinskianus',
                             species == 'HETARB' ~ 'H. arbutifolia',
                             species == 'MALLAU' ~ 'M. laurina',
                             species == 'SALAPI' ~ 'S. apiana',
                             species == 'SALLEU' ~ 'S. leucophylla'))

# color
r.flamm.index.top$color <- c('#00468BB2', '#ED0000B2', '#42B540B2', '#0099B4B2', '#925E9FB2', '#AD002AB2', '#ADB6B6B2', '#1B1919B2')

# legend triangle
r.flamm.index.legend <- data.frame(species = c('Legend'), ignitability = c(0.5), combustibility = c(0.5), sustainability = c(0.5), color = c('white'))
```

Plotting
```{r}
par(mfrow = c(3, 3), mar = c(0, 0, 2, 0.01))

# Species Triangles
for(i in 1:nrow(r.flamm.index.top)){
TernaryPlot(" ", " ", " ",
            grid.lines = 0,
            grid.minor.lines = 0,
            lab.cex = 3.1,
            bg = 'transparent',
            col = r.flamm.index.top$color[i])

title(r.flamm.index.top[i,1], cex.main = 3.5, font.main = 4)


# Add an arrow
TernaryArrows(c(50, 50, 50), c(r.flamm.index.top$ignitability[i], 0.15, 0.15), length = 0.1*r.flamm.index.top$ignitability[i], lwd = 2.3*r.flamm.index.top$ignitability[i], col = "black")
TernaryArrows(c(50, 50, 50), c(0.15, r.flamm.index.top$combustibility[i], 0.15), length = 0.1*r.flamm.index.top$combustibility[i], lwd = 2.3*r.flamm.index.top$combustibility[i], col = "black")
TernaryArrows(c(50, 50, 50), c(0.15, 0.15, r.flamm.index.top$sustainability[i]), length = 0.1*r.flamm.index.top$sustainability[i], lwd = 2.3*r.flamm.index.top$sustainability[i], col = "black")
}

# Legend Triangle
TernaryPlot("Ign.", "Comb.", "Sus.",
            grid.lines = 0,
            grid.minor.lines = 0,
            lab.cex = 3.1,
            bg = 'transparent',
            col = r.flamm.index.legend$color[1])

title(r.flamm.index.legend$species[1], cex.main = 3.5)

# Add an arrow
TernaryArrows(c(50, 50, 50), c(r.flamm.index.legend$ignitability[1], 0.15, 0.15), length = 0.1*r.flamm.index.legend$ignitability[1], lwd = 2.3*r.flamm.index.legend$ignitability[1], col = "black")
TernaryArrows(c(50, 50, 50), c(0.15, r.flamm.index.legend$combustibility[1], 0.15), length = 0.1*r.flamm.index.legend$combustibility[1], lwd = 2.3*r.flamm.index.legend$combustibility[1], col = "black")
TernaryArrows(c(50, 50, 50), c(0.15, 0.15, r.flamm.index.legend$sustainability[1]), length = 0.1*r.flamm.index.legend$sustainability[1], lwd = 2.3*r.flamm.index.legend$sustainability[1], col = "black")

flam.triangles <- recordPlot()
```

## Graphical Abstract 
Two Fire Triangles
```{r}
par(mfrow = c(1, 2), mar = c(0, 0, 1, 0.01))
for(i in c(4,8)){
TernaryPlot("Ignitability", "Combustibility", "Sustainabilty",
            grid.lines = 0,
            grid.minor.lines = 0,
            lab.cex = 1.5,
            bg = 'transparent',
            col = r.flamm.index.top$color[i])

# Add an arrow
TernaryArrows(c(50, 50, 50), c(r.flamm.index.top$ignitability[i], 0.15, 0.15), length = 0.1*r.flamm.index.top$ignitability[i], lwd = 2*r.flamm.index.top$ignitability[i], col = "black")
TernaryArrows(c(50, 50, 50), c(0.15, r.flamm.index.top$combustibility[i], 0.15), length = 0.1*r.flamm.index.top$combustibility[i], lwd = 2*r.flamm.index.top$combustibility[i], col = "black")
TernaryArrows(c(50, 50, 50), c(0.15, 0.15, r.flamm.index.top$sustainability[i]), length = 0.1*r.flamm.index.top$sustainability[i], lwd = 2*r.flamm.index.top$sustainability[i], col = "black")
}
flam.triangles.graphical <- recordPlot()
```

# Flam. vs. Desirable Traits (Scores)
## Data Wrangling
```{r}
# calculating desirable traits scores
## The below numbers are average 'scores' assigned by experts describing how well each of the study species aligns with identified desirable plant traits
flam_EScores_subscores <- data.frame(
  Species = c('ARTCAL', 'ARCDEN', 'CEAGRI', 'ERIKAR', 'HETARB', 'MALLAU', 'SALAPI', 'SALLEU'),
  drought =  c(3.33, 3.5, 3.33, 3, 3.5, 3.6, 3.5, 3.5), # drought tolerance
  edible =   c(2.5, 2, 1.33, 2.4, 2.166, 2, 2.5, 2.33),
  flowers =  c(3.33, 2.0833, 3.5, 3.4, 2.9166, 2.7, 3.33, 3.5833),
  fragrant = c(2.833, 4, 2.833, 1.8, 2.33, 2.8, 4, 4),
  wildlife = c(3.5, 3.5, 3.33, 1.8, 3.666, 3.6, 3.333, 3.333), # wildlife benefits
  erosion = c(3.833, 4, 4, 2.6, 4, 4, 3.833, 3.833), # erosion control
  flowers.aes = c(3.33, 2.166, 3.66, 4, 2.66, 2.6, 3.5, 3.833), # aesthetics of flowers
  flowers.pol = c(3.33, 2, 3.33, 2.8, 3.166, 2.8, 3.166, 3.33) # benefits of flowers to pollinators
)  

flam_EScores_subscores$ES.score <- (0.556*flam_EScores_subscores$drought + 0.222*flam_EScores_subscores$edible + 0.074*flam_EScores_subscores$flowers + 0.037*flam_EScores_subscores$fragrant + 0.111*flam_EScores_subscores$wildlife) # the following score was calculated based on the % of survey responses that highly rated a given desirable trait

# scores DF
flam_scores <- data.frame(
  species = flam_EScores_subscores$Species,
  flam.score = r.flamm.index.top$flam.score, # using flammability scores calculated from above
  es.score = flam_EScores_subscores$ES.score
)

# scale flammability and desirability scores
flam_scores <- flam_scores %>% 
  mutate(flam.score = scale(flam.score), scaled.es.score = scale(es.score))
```

## Visualization
For this version, top-performing models were used and fixed and random effects were removed
*This is the version we preferred*
```{r}
# Visualization
# 'ARTCAL', 'ARCDEN', 'CEAGRI', 'ERIKAR', 'HETARB', 'MALLAU', 'SALAPI', 'SALLEU'
flam_scores$labx <- c(0.06, -0.31, -1.4, -2.08, 0.32, 0.38, 0.73, 0.57)
flam_scores$laby <- c(0.02, -0.4, 1.38, -1.32, -0.65, -0.33, 0.13, 1.8)

es.vs.flam <- ggplot(data=flam_scores) +
  geom_point(aes(x=scaled.es.score, y=flam.score, color=species), size=12, alpha=0.75) + 
  geom_text(vjust=0, hjust=-0.15, aes(x=labx, y=laby, label=species), fontface = 'bold', size = 12) +
  annotate(geom = 'segment', x = -1.5, xend = 1, y = -2, yend = -2, linewidth = 2,
           arrow = arrow(ends = 'both', angle = 20), color = 'gray60') + # desirability score arrow
  annotate('text', x = 1.22, y = -2, label = 'More \nDesirable', size = 9,
           fontface = 'bold', color = 'gray60') +
  annotate('text', x = -1.72, y = -2, label = 'Less \nDesirable', size = 9,
           fontface = 'bold', color = 'gray60') +
  annotate(geom = 'segment', x = -2.16, xend = -2.16, y = -1.35, yend = 1.75, linewidth = 2,
           arrow = arrow(ends = 'both', angle = 20), color = 'gray60') + # flammability score arrow
  annotate('text', x = -2.13, y = 1.97, label = 'More \nFlammable', size = 9,
           fontface = 'bold', color = 'gray60') +
  annotate('text', x = -2.13, y = -1.57, label = 'Less \nFlammable', size = 9,
           fontface = 'bold', color = 'gray60') +
  labs(x = 'Desirable Traits Score', y = 'Flammability Score') +
  scale_color_manual(values = c('#00468BB2', '#ED0000B2', '#42B540B2', '#0099B4B2', '#925E9FB2', '#AD002AB2', '#ADB6B6B2', '#1B1919B2')) +
  theme_bw() +
  scale_y_continuous(limits = c(-2, 2)) +
  scale_x_continuous(limits = c(-2.22, 1.3)) +
  theme(legend.position = 'none',
          axis.title = element_text(face = 'bold', size = 36),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          panel.grid = element_blank())
          #axis.text = element_text(size = 22))
es.vs.flam
```

## Updating Visualization
```{r}
# Visualization
# 'ARTCAL', 'ARCDEN', 'CEAGRI', 'ERIKAR', 'HETARB', 'MALLAU', 'SALAPI', 'SALLEU'
flam_scores$laby_updated <- c(-1.06, -0.62, 0.78, -1.32, -0.25, 0.03, 0.45, 1.8)
flam_scores$labx_updated <- c(3.1, 2.55, 2.59, 2.47, 3.15, 3.16, 3.2, 2.77)
flam_scores$species_updated <- c('A. cal.', 'A. den.', 'C. gri.', 'E. kar.', 'H. arb.', 'M. lau.', 'S. api.', 'S. leu.')

es.vs.flam <- ggplot(data=flam_scores) +
  geom_point(aes(x=es.score, y=flam.score, color=species), size=12, alpha=0.75) + 
  geom_text(vjust=0, hjust=-0.15, aes(x=labx_updated, y=laby_updated, label=species_updated), fontface = 'bold.italic', size = 12) +
  #annotate(geom = 'segment', x = 1.1, xend = 1.1, y = -1.65, yend = 1.65, linewidth = 2,
  #         arrow = arrow(ends = 'both', angle = 20), color = 'gray60') + # flammability score arrow
  #annotate('text', x = 1.1, y = 1.93, label = 'More \nFlammable', size = 9,
  #         fontface = 'bold', color = 'gray60') +
  #annotate('text', x = 1.1, y = -1.93, label = 'Less \nFlammable', size = 9,
  #         fontface = 'bold', color = 'gray60') +
  labs(x = 'Desirable Traits Score', y = 'Flammability Score') +
  scale_color_manual(values = c('#00468BB2', '#ED0000B2', '#42B540B2', '#0099B4B2', '#925E9FB2', '#AD002AB2', '#ADB6B6B2', '#1B1919B2')) +
  theme_bw() +
  scale_y_continuous(limits = c(-2, 2)) +
  scale_x_continuous(limits = c(1, 4)) +
  theme(legend.position = 'none',
          axis.title = element_text(face = 'bold', size = 36),
          #axis.text.y = element_blank(),
          axis.ticks = element_blank(),
          panel.grid = element_blank(),
          axis.text = element_text(size = 36))
es.vs.flam
```

### Stats
```{r}
# linear regression
desireflam.lm <- lm(flam.score ~ es.score, data = flam_scores)
summary(desireflam.lm)

```


## Adding Defensible Space Zones
```{r}
es.vs.flam.zones <- ggplot(data=flam_scores) +
  geom_point(aes(x=es.score, y=flam.score, color=species), size=12, alpha=0.75) + 
  geom_text(vjust=0, hjust=-0.15, aes(x=labx_updated, y=laby_updated, label=species_updated), fontface = 'bold.italic', size = 12) +
  #annotate(geom = 'rect', xmin = 1, xmax = 4, ymin = -2, ymax = -1,
  #         alpha = 0.2, fill = '#00CC99') +
  #annotate('text', x = 1.1, y = -1.5, label = 'Zone 1', size = 9,
  #         fontface = 'bold', color = '#005641', angle = 90) +
  #annotate(geom = 'rect', xmin = 1, xmax = 4, ymin = -1, ymax = 1,
  #         alpha = 0.2, fill = '#FFEE88') +
  #annotate('text', x = 1.1, y = 0, label = 'Zone 2', size = 9,
   #        fontface = 'bold', color = '#AE9400', angle = 90) +
  #annotate(geom = 'rect', xmin = 1, xmax = 4, ymin = 1, ymax = 2,
  #         alpha = 0.2, fill = '#D7263D') +
  #annotate('text', x = 1.1, y = 1.5, label = 'Avoid', size = 9,
  #         fontface = 'bold', color = '#750010', angle = 90) +
  labs(x = 'Desirable Traits Score', y = 'Flammability Score') +
  scale_color_manual(values = c('#00468BB2', '#ED0000B2', '#42B540B2', '#0099B4B2', '#925E9FB2', '#AD002AB2', '#ADB6B6B2', '#1B1919B2')) +
  theme_bw() +
  scale_y_continuous(limits = c(-2, 2)) +
  scale_x_continuous(limits = c(1, 4)) +
  theme(legend.position = 'none',
          axis.title = element_text(face = 'bold', size = 36),
          #axis.text.y = element_blank(),
          axis.ticks = element_blank(),
          panel.grid = element_blank(),
          axis.text = element_text(size = 36))
es.vs.flam.zones
```


# Main Figure
Arranging Fire Triangles and Scores Scatterplot and Saving
```{r}
plot_grid(es.vs.flam.zones, flam.triangles, labels = c('a', 'b'), nrow = 1, rel_widths = c(2.5, 2), label_size = 50, greedy = F)
ggsave(file = here('figures', 'main_figures', 'Fig7_conceptual_plot.png'), height = 10, width = 25)
```
